% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddEcoCrop.R
\name{AddEcoCrop}
\alias{AddEcoCrop}
\title{Get EcoCrop parameters for ERA products}
\usage{
AddEcoCrop(Products)
}
\arguments{
\item{Products}{A character vector of products that match the \code{Product} field of the \code{EUCodes} dataset}
}
\value{
A data.table is returned with fields:
\itemize{
\item \code{species} = scientific name of crop (a unique identifier in the \code{ERA_EcoCrop} dataset)
\item \code{Topt.low} = minimum of crop optimal temperature range - C
\item \code{Topt.high} = maximum of crop optimal temperature range - C
\item \code{Tlow} = minimum of crop absolute temperature range - C
\item \code{Thigh} = maximum of crop absolute temperature range - C
\item \code{cycle_min} = minimum of crop growing cycle duration range - days
\item \code{cycle_max} = maximum of crop growing cycle duration range - days
}
}
\description{
Extracts ecocrop temperature and cycle length variables for a vector of product codes.
}
