% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcClimate.R
\name{CalcClimate}
\alias{CalcClimate}
\title{Calculate climate statistics}
\usage{
CalcClimate(
  DATA,
  CLIMATE,
  ID,
  Rain.Data.Name,
  Temp.Data.Name,
  Rain.Windows = c(6 * 7, 4 * 7, 2 * 7, 2 * 7),
  Widths = c(3, 3, 2, 2),
  Rain.Threshold = c(30, 30, 20, 15),
  Do.LT.Avg = T,
  Max.LT.Avg = 2010,
  Do.BioClim = T,
  Windows = data.table(Name = "Plant.0-30", Start = 1, End = 30),
  SaveDir = "Climate Stats/",
  ErrorDir = "Climate Stats/Errors/",
  ROUND = 5
)
}
\arguments{
\item{DATA}{An ERA dataset (e.g. \code{ERA.Compiled}) processed by the \code{AddEcoCrop}, \code{EstPDayData}, \code{EstSLenData} and \code{EstPDayRain} functions.}

\item{CLIMATE}{A daily agroclimatology dataset with fields: 1) \code{Temp.Mean} = mean daily temperature - C; 2) \code{Temp.Max} = maximum daily temperature - C;
\code{Temp.Min} = minimum daily temperature - C; 4) \code{Rain} = daily rainfall - mm; 5) \code{ETo} = reference evapotranspiration - mm; 6) class \code{Date} field
named \code{Date}; and 7) location identity as per \code{ID} argument and named the same as per \code{DATA}.}

\item{ID}{A character vector of length one containing the column name for a unique id field naming each location in the dataset provided}

\item{Rain.Data.Name}{A character vector of length one containing the name of the rainfall dataset in \code{CLIMATE}, this must be the same has the rainfall data used when preparing \code{DATA} using
the \code{EstPDayRain} function. This can be the same as \code{Temp.Data.Name}.}

\item{Temp.Data.Name}{A character vector of length one containing the name of the agroclimatology dataset in \code{CLIMATE}.}

\item{Rain.Windows}{An integer vector; the width in days of four temporal windows within which rainfall is assessed for potential planting dates. The first
value in the series specifies a period before the planting date, then subsequent values define three consecutive windows after the planting date.}

\item{Widths}{An integer vector of length equivalent to length(\code{Window}); the width of the scanning window in days within which rainfall is summed for
the corresponding \code{Rain.Windows} entry.}

\item{Rain.Threshold}{An integer vector of length equivalent to length(\code{Rain.Windows}); the amount of rainfall that has to fall in the temporal windows considered.}

\item{Do.LT.Avg}{Logical \code{T/F;} if \code{T} long term averages and deviances are calculated.}

\item{Max.LT.Avg}{Integer of length one only relevant if \code{Do.LT.Avg==T}; maximum year considered when calculating long-term averages.}

\item{Do.BioClim}{Logical \code{T/F}; if \code{T} calculate annual and long-term average bioclimatic variables from the agroclimatic data provided in \code{CLIMATE} using the \link[dismo]{biovars} function}

\item{Windows}{A data.table with three columns \code{Name}, \code{Start} and \code{End} which specifies additional temporal periods for calculation of climate statistics.
For example \code{data.table(Name="Plant.1-30",Start=1,End=30)} is a temporal period from the day after planting to 30 days after planting. Set to NA if no additional windows are required.}

\item{SaveDir}{A character vector of length one containing the path to the directory where the output is saved. Set to NA if you do not want to save the returned datasets.}

\item{ErrorDir}{A character vector of length one containing the path to the directory where information on potential analysis errors is to be saved. Set to NA if you do not want to save the returned dataset.}

\item{ROUND}{An integer vector of length one indicating the number of decimal places to round output values to.}
}
\value{
A list is output containing following data.tables:
\strong{\verb{[[Observed]]}} = A \code{data.table} of seasonal climate statistics for
\strong{\verb{[[LongTerm]]}} =
\emph{\verb{[[LongTerm]][[LT.PD.Years]]}} =
\emph{\verb{[[LongTerm]][[LT.PD.Avg]]}} =
\emph{\verb{[[LongTerm]][[LT.Clim.Years]]}} =
*\emph{\verb{[[LongTerm]][[LT.Clim.Avg]]}} =
\strong{\verb{[[Annual.BioClim]]}} =
\emph{\verb{[[Annual.BioClim]][[Annual.Estimates]]}} =
\emph{\verb{[[Annual.BioClim]][[LT.Averages]]}} =
\emph{\verb{[[Annual.BioClim]][[Key]]}} =
\emph{\verb{[[Parameters]]}} = List of argument values supplied to the function
}
\description{
The \code{CalcClimate} function
}
\details{
Process:
\enumerate{
\item Season Length \code{SLen} is calculated as difference between mean planting and harvest dates.
\item The \code{SLen} function is applied to the data which combines accurate temporal planting/harvest data with estimates using the following and sequential logic:
i) where \code{(Plant.End-Plant.Start)>20}and a planting date derived from the \code{EstPDayRain} function is available \code{SLen} is substituted for
\code{Harvest.End} - rainfall estimate planting date;
ii) where \code{(Plant.End-Plant.Start<20) & (Harvest.End-Harvest.Start>=20)} \code{SLen} is substituted for \code{Harvest.End-(Plant.Start+(Plant.End-Plant.Start)/2)};
iii) A \code{P.Date.Merge} field is added to \code{DATA} as \code{Plant.Start+(Plant.End-Plant.Start)/2};
iv) Missing \code{NA} values of \code{P.Date.Merge} are substituted with estimates derived from similar nearby observations estimated using the \code{EstPDayData}
function;
v) A \code{SLen.Merge}  field is added to \code{DATA} equivalent to the \code{SLen} field with missing \code{NA} values substituted with estimates derived from similar nearby
observations estimated using the \code{EstSLenData} function;
vi) A \code{SLen.Ecocrop} field is added to \code{DATA} as \code{(cycle_min+cycle_max)/2};
vii) \code{DATA} is subset to data.table named \code{SS} which contains unique values of location, product, observation date and planting/harvest/season length
fields. Some form of planting date and season length estimate must be present;
vii) Where there is a difference of greater than 60\% between ecocrop vs data estimates of season length observations are rejected.
\emph{The use of Harvest.End in Steps i & ii is deliberate; when uncertainty is large using mean planting and harvest dates can result in unrealistically
short season lengths that distort measures of seasonal climate, especially precipitation. The use of Harvest.End may overestimate season length overall for highly uncertain
scenarios, but we consider the effect this on seasonal climate measures less severe.}
\item For each row of \code{SS} the \code{CLIMATE} dataset is filtered on the \code{Date} field for two temporal periods based on the date fields \code{P.Date.Merge} and \code{SLen.Merge}
or \code{SLen.Ecocrop}, note one day is added to \code{P.Date.Merge} so the temporal window begins the day after planting. An additional temporal windows can
be specified using data.table supplied to the \code{Windows} argument, the data.table requires three fields: 1) \code{Name}; 2) \code{Start} and 3) \code{End}. The
\code{Start} and \code{End} fields define the additional temporal period as days after\code{P.Date.Merge}.For example \code{Window=data.table(Name="Plant.1-30",Start=1,End=30)}
is a temporal period from planting to 30 days after planting.
\item For each temporal period specified in 3) climate statistics are calculated:
i) \strong{growing degrees days (GDD)} are calculated using the \code{GDDcalc} function which requires daily \code{Tmax} and \code{Tmin} fields from \code{CLIMATE} and the EcoCrop
optimal and absolute temperate thresholds (fields \code{Tlow}, \code{Topt.low}, \code{Topt.high}, and \code{Thigh}) added to \code{DATA} using the \code{AddEcoCrop} function.
The \code{GDDcalc} function outputs four fields \code{GDDlow}, \code{GDDopt}, \code{GDDhigh} and \code{GDDmax} and the values of these are summed for the temporal period;
ii) \strong{precipitation and reference evapotranspiration (ETo)} statistics are estimated using the \code{RAIN.Calc} function which outputs a number of variables
described in the documentation for this function;
iii) \strong{max and mean temperatures} are averaged with standard deviation (\code{Tmax.Mean},\code{Tmax.sd},\code{Tmean.mean} and \code{Tmean.sd} fields).
\item If argument \code{Do.LT.Avg==T} annual and long-term climate statistics an planting dates are calculated  for each combination of location and product. Note
that each season in the data is treated separately. The process logic is as follows:
i) The median planting julian day of the year is taken for the location x product combination;
ii) For each year of complete climate data, the median planting date from i) is used as a starting point to estimate planting date from rainfall using the
\code{Est.Rain} function which takes the arguments \code{Rain.Windows}, \code{Widths} and \code{Rain.Threshold} as explained in the arguments for this function. These data are
returned as \verb{[[LongTerm]][[LT.PD.Years]]};
iii) Estimated planting dates from ii) are averaged (mean, sd, median) across years and returned as \verb{[[LongTerm]][[LT.PD.Avg]]};
iv) Annual planting date deviance is calculated as the absolute difference between the annual and mean/median planting dates and appended to
the output of ii);
v) Annual climate statistics are calculated for each year x site x product combination using the rainfall estimated planting date and as per steps
3 & 4. Data are return as \verb{[[LongTerm]][[LT.Clim.Years]]}. Where insufficient rainfall fell to meet the thresholds specified in ii) then the a median
planting date is substituted from those years that did have sufficient rainfall in ii);
vi) Long-term climate statistics from v) are calculated across years for mean, median, standard deviation, minimum and maximum statistics. Data are
returned \verb{[[LongTerm]][[LT.Clim.Avg]]};
vii) Deviance from long-term mean and median values for each climate statistic is appended as columns to the output of v).
\item If argument \code{Do.BioClim==T} then:
i) Annual bioclimatic values are calculated using the \link[dismo]{biovars} function for each complete year of
data in \code{CLIMATE} per unique locationas (\code{ID} field) in \code{DATA}. The output data is return as [\link{Annual.BioClim}][\link{Annual.Estimates}]\verb{;   ii) Bioclim variables are averaged over time to give long-term mean and median values. The output data is return as [[Annual.BioClim]][[LT.Averages]]};
iii) Deviance from long-term mean and median values for each climate statistic is appended as columns to the output of i).
}
}
