% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyNegVals.R
\name{ClassifyNegVals}
\alias{ClassifyNegVals}
\title{Classify outcomes with negative numbers that are incompatible with ratio approaches}
\usage{
ClassifyNegVals(
  Data,
  OCode,
  Thresholds = c(0.7, 0.95, 1.05, 1.3),
  Vals = c(`0` = 0, `++` = 2, `+` = 1, `-` = -1, `--` = -2, `+/-` = 1, `-/+` = -1, `+0`
    = 1, `0+` = -1, `-0` = -1, `0-` = 1),
  Invert2xNeg = T
)
}
\arguments{
\item{Data}{An ERA data.table (e.g. \code{ERAg::ERA.Compiled}).}

\item{OCode}{An ERA outcome code (see \code{ERAg::OutcomeCodes[,Codes]}.}

\item{Thresholds}{A numeric vector of length four corresponding to classification thresholds for worst, poor, good & best outcomes. If:
\itemize{
\item MeanT/MeanC<Threshold\link{1} then Class = "--"
\item Threshold\link{1}<=MeanT/MeanC<Threshold\link{2} then Class = "-"
\item Threshold\link{2}<=MeanT/MeanC<Threshold\link{3} then Class = "0"
\item Threshold\link{3}<=MeanT/MeanC<Threshold\link{4} then Class = "+"
\item MeanT/MeanC>=Threshold\link{4} then Class = "++"
}}

\item{Vals}{A numeric vector of length 11 proving numeric scores for each outcome classification:
\itemize{
\item Vals\link{1}: Class "0" = \code{MeanT/MeanC<Threshold[1]}
\item Vals\link{2}: Class "++" = \verb{Threshold[1]<=MeanT/MeanC<Threshold[2]}
\item Vals\link{3}: Class "+"= \verb{Threshold[2]<=MeanT/MeanC<Threshold[3]}
\item Vals\link{4}: Class "-" = \verb{Threshold[3]<=MeanT/MeanC<Threshold[4]}
\item Vals\link{5}: Class "--" = \code{MeanT/MeanC>=Threshold[4]}
\item Vals\link{6}: Class "+/-" = \code{MeanT>0 & MeanC<0}
\item Vals\link{7}: Class "-/+" =\code{ MeanT<0 & MeanC>0}
\item Vals\link{8}: Class "+0" = \code{MeanT>0 & MeanC==0}
\item Vals\link{9}: Class "0+" = \code{MeanT==0 & MeanC>0}
\item Vals\link{10}: Class "-0" = \code{MeanT<0 & MeanC==0}
\item Vals\link{11}: Class "0-" = \code{MeanT==0 & MeanC<0}
}}

\item{Invert2xNeg}{Logical \code{T/F}. Where MeanC<0 & MeanT<0 invert the calculation of yi to \code{log(MeanC/MeanT)}}
}
\value{
ClassifyNegVals returns the input Data subset to the outcome provided in the \code{OCode}argument and with the additional columns:
\itemize{
\item \code{Class} a character vector of class names
\item \code{Class.Val} a numeric vector of class values
}
}
\description{
For a the supplied outcome argument \code{OCode} this function classifies the difference between \code{MeanC} and \code{MeanT} into character and numeric columns based on the
\code{Threshold} and \code{Vals} arguments. These columns are appended to the input dataset that has been subset to the outcome of interest.
}
\details{
When generating ratios the function will invert outcomes where \code{MeanT} and \code{MeanC} are both negative unless the \code{Invert2xNeg} argument is set to \code{FALSE}.
}
