% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConceptLevels.R
\name{CreateConceptLevels}
\alias{CreateConceptLevels}
\title{ERA Concept Hierarchy Tables}
\usage{
CreateConceptLevels()
}
\value{
CreateConceptLevels returns a list of four elements each a \code{data.table}:
\itemize{
\item \code{Prac.Levels} Practice concept organizational levels
\item \code{Out.Levels} Outcome concept organizational levels
\item \code{Prod.Levels} Product (also called EU or experimental units) concept organizational levels
\item \code{Agg.Levels} This data.table describes fields that can be used to aggregate data. \code{Observation} = no aggregation; each row of ERA is an observation, \code{Study} = the publication/study an
observation is reported from, and \code{Location} the spatial location of an observation (a combination of latitude, longitude and a buffer of uncertainty).
\code{Choice} and \code{Choice.Code} fields are simplified names/codes for hierarchical concept levels for user selection task. The remaining columns indicate the corresponding
column names in the ERA dataset for user selections.
}
}
\description{
Creates a list of data.tables describing the organizational structure of the ERA concept hierarchy for use in other ERA functions and interactive apps.
}
