% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_outcome.R
\name{derive_outcome}
\alias{derive_outcome}
\title{derive_outcome
Calculates new outcomes by dividing or subtracting one outcome from another. If function returns NULL then no derived data are present. Any observation outcome sets that already have a result outcome associated with them are removed.}
\usage{
derive_outcome(
  Data,
  outcome_col = "Out.SubInd",
  outcome1,
  outcome2,
  outcome_result,
  operation = "subtract",
  match_fields,
  val_col1 = "MeanT",
  val_col2 = "MeanC"
)
}
\arguments{
\item{Data}{An ERA data.table (e.g. \code{ERAg::ERA.Compiled}).}
}
\description{
derive_outcome
Calculates new outcomes by dividing or subtracting one outcome from another. If function returns NULL then no derived data are present. Any observation outcome sets that already have a result outcome associated with them are removed.
}
