% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_outcome_sub.R
\name{derive_outcome_sub}
\alias{derive_outcome_sub}
\title{derive_outcome_sub
Calculates new outcomes by subtracting one outcome from another. If function returns NULL then no derived data are present. Any observation outcome sets that already have a result outcome associated with them are removed.}
\usage{
derive_outcome_sub(Data, target = 120, subtract = 150, result = 124)
}
\arguments{
\item{Data}{An ERA data.table (e.g. \code{ERAg::ERA.Compiled}).}

\item{target}{ERA outcome code of target outcome (see \code{ERAg::OutcomeCodes})}

\item{subtract}{ERA outcome code of outcome that will be substracted from the target outcome (see \code{ERAg::OutcomeCodes})}

\item{result}{ERA outcome code for the resulting values (see \code{ERAg::OutcomeCodes})}
}
\description{
derive_outcome_sub
Calculates new outcomes by subtracting one outcome from another. If function returns NULL then no derived data are present. Any observation outcome sets that already have a result outcome associated with them are removed.
}
