% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadCHIRPS.R
\name{DownloadCHIRPs}
\alias{DownloadCHIRPs}
\title{Download CHIRPS 2.0 dataset}
\usage{
DownloadCHIRPs(
  StartYear = 1983,
  EndYear = 2020,
  SaveDir = paste0(getwd(), "/CHIRPS")
)
}
\arguments{
\item{StartYear}{A single integer value from 1983 onward indicating the start year for data download. Default value = \code{1981}.}

\item{EndYear}{A single integer value from 1983 onward indicating the end year for data download. Default value = \code{2020}.}

\item{SaveDir}{A character vector of length one containing the path to the directory where CHIRPS data should save. Default value = \verb{/CHIRPS}.}
}
\description{
This function download the CHIRPS 2.0 dataset at 0.05 degree resolution from the repository https://data.chc.ucsb.edu/products/CHIRPS-2.0/africa_daily/tifs/p05/1981/ to
a local directory specied by the \code{SaveDir} argument.
}
