% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA.Compiled}
\alias{ERA.Compiled}
\title{The compiled ERA dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 112621 rows and 141 columns.
}
\source{
\emph{To be added}
}
\usage{
ERA.Compiled
}
\description{
\describe{
\item{Index}{unique row identity}
\item{Code}{unique publication identity}
\item{Author}{primary author of the publication}
...
}
}
\keyword{datasets}
