% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_AEZ_MAP_MAT}
\alias{ERA_AEZ_MAP_MAT}
\title{Agroecological zones}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1617 rows and 6 columns.
}
\source{
HarvestChoice; International Food Policy Research Institute (IFPRI), 2015, "Agro-Ecological Zones for Africa South of the Sahara", https://doi.org/10.7910/DVN/M7XIUB, Harvard Dataverse, V3
}
\usage{
ERA_AEZ_MAP_MAT
}
\description{
\describe{For each of the unique ERA locations and its buffer of spatial uncertainty, the agroecological zone is determined based on one of the three
classification schemes
\item{Site.Key}{longitude, latitude and buffer for the unique ERA location}
\item{AEZ16simple}{shorter name for AEZ16}
\item{AEZ16}{agroecological zone using the scheme with 16 classes}
\item{AEZ5}{agroecological zone using the scheme with 5 classes}
\item{Mean.Annual.Precip}{mean annual precipitation at the ERA location}
\item{Mean.Annual.Temp}{mean annual precipitation at the ERA location}
...
}
}
\keyword{datasets}
