% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_Bibliography}
\alias{ERA_Bibliography}
\title{ERA_Bibliography}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2147 rows and 16 columns.
}
\source{
\emph{To be added}
}
\usage{
ERA_Bibliography
}
\description{
\describe{The fields used to collect details of the publication used to collect ERA data.
\item{CATEGORY}{the type of publication}
\item{AUTHOR}{authors of the publication}
\item{BOOKTITLE}{title of the book. NA if the publication is in a journal}
\item{JOURNAL}{journal in which the article was published}
\item{NUMBER}{journal number, which refers to how many times that periodical has been published during that year}
\item{PAGES}{pages of the journal or book}
\item{TITLE}{the title of the article}
\item{TYPE}{type of article}
\item{VOLUME}{the volumn of a journal}
\item{YEAR}{the year of publication}
\item{DOI}{doi of the article. DOI = Digital Object Identifier}
\item{ISSN}{issn of the article. ISNN = International Standard Serial Number}
\item{URL}{url of the article. The URL provides the wed address of the journal. URL = Uniform Resource Locator}
\item{ERACODE}{unique data entry code}
\item{KEYWORDS}{keywords used in the article}
\item{ABSTRACT}{a brief summary of the publication, collected from the article}
...
}
}
\keyword{datasets}
