% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_BioClim}
\alias{ERA_BioClim}
\title{Bioclimatic Variables}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1533 rows and 115 columns.
}
\source{
https://www.worldclim.org/data/bioclim.html version 2.1 climate data for 1970-2000, released in January 2020.
}
\usage{
ERA_BioClim
}
\description{
\describe{There are  19 bioclimatic variables which are derived from the monthly temperature and rainfall values in order to generate biologically meaningful variables.
These are often used in species distribution modeling and related ecological modeling techniques. The bioclimatic variables represent annual
trends (e.g., mean annual temperature, annual precipitation) seasonality (e.g., annual range in temperature and precipitation) and extreme or
limiting environmental factors (e.g., temperature of the coldest and warmest month, and precipitation of the wet and dry quarters). A quarter
is a period of three months (1/4 of the year).
Bioclim data are extracted and summarized for each unique ERA location plus its buffer of spatial uncertainty.
\item{bio_1}{Annual Mean Temperature}
\item{bio_2}{Mean Diurnal Range (Mean of monthly (max temp - min temp))}
\item{bio_3}{Isothermality (BIO2/BIO7) (* 100)}
\item{bio_4}{Temperature Seasonality (standard deviation *100)}
\item{bio_5}{Max Temperature of Warmest Month}
\item{bio_6}{Min Temperature of Coldest Month}
\item{bio_7}{Temperature Annual Range (BIO5-BIO6)}
\item{bio_8}{Mean Temperature of Wettest Quarter}
\item{bio_9}{Mean Temperature of Driest Quarter}
\item{bio_10}{Mean Temperature of Warmest Quarter}
\item{bio_11}{Mean Temperature of Coldest Quarter}
\item{bio_12}{Annual Precipitation}
\item{bio_13}{Precipitation of Wettest Month}
\item{bio_14}{Precipitation of Driest Month}
\item{bio_15}{Precipitation Seasonality (Coefficient of Variation)}
\item{bio_16}{Precipitation of Wettest Quarter}
\item{bio_17}{Precipitation of Driest Quarter}
\item{bio_18}{Precipitation of Warmest Quarter}
\item{bio_19}{Precipitation of Coldest Quarter}
...
}
}
\keyword{datasets}
