% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_CCI_LC_15}
\alias{ERA_CCI_LC_15}
\title{Landcover}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1520 rows and 31 columns.
}
\source{
CCI-LC Land Cover Maps - v2.0.7 http://maps.elie.ucl.ac.be/CCI/viewer/download.php
}
\usage{
ERA_CCI_LC_15
}
\description{
\describe{
The CCI-LC project delivers a new time series of 24 consistent global LC maps at 300 m spatial resolution on an annual basis from 1992 to 2015.
The number of raster cells of each landcover class for the year 2015 is summed for each unique ERA location and its buffer of spatial uncertainty.
A description of the landcover classes (fields) in this dataset can be found in the \emph{\code{ERA_CCI_LC_15_Fields}} object.
...
}
}
\keyword{datasets}
