% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_ISDA}
\alias{ERA_ISDA}
\title{ISDA Soil Parameters}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 48741 rows and 9 columns.
}
\source{
ISDA: https://www.isda-africa.com/isdasoil/  African Soil and Agronomy Data Cube: https://gitlab.com/openlandmap/africa-soil-and-agronomy-data-cube.
}
\usage{
ERA_ISDA
}
\description{

}
\details{

}
\keyword{datasets}
