% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_Physical}
\alias{ERA_Physical}
\title{ERA_Physical}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1514 rows and 20 columns.
}
\source{
Void filled Aster GDEM version 2 downloaded from http://www.viewfinderpanoramas.org/Coverage\%20map\%20viewfinderpanoramas_org3.htm
}
\usage{
ERA_Physical
}
\description{
\describe{
\item{Latitude}{latitude of study location in decimal degrees}
\item{Longitude}{longitude of study location in decimal degrees}
\item{Buffer}{estimation in m of the spatial uncertainty for a point location}
\item{Site.Key}{longitude, latitude and buffer for the site}
\item{Country}{country in which the site is located in}
\item{ISO.3166.1.alpha.3}{three-letter country codes. See: https://www.iso.org/obp/ui/#search}
\item{DEMcells}{number of DEM cells within the site buffer}
\item{Altitude.med}{median altitude of the site}
\item{Altitude.mean}{mean altitude of the site}
\item{Altitude.sd}{standard deviation of altitude at the site}
\item{Altitude.max}{maximum altitude of the site}
\item{Altitude.min}{minimum altitude of the site}
\item{Slope.med}{median slope of the site}
\item{Slope.mean}{mean slope of the site}
\item{Slope.sd}{standard deviation of slope at the site}
\item{Slope.max}{maximum slope of the site}
\item{Slope.min}{minimum slope of the site}
\item{Aspect.med}{median aspect of the site}
\item{Aspect.mean}{mean aspect of the site}
\item{Aspect.sd}{standard deviation of aspect at the site}
...
}
}
\keyword{datasets}
