% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_Search_Terms}
\alias{ERA_Search_Terms}
\title{ERA_Search_Terms}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 19 rows and 3 columns.
}
\source{
ERA Project Team (2021)
}
\usage{
ERA_Search_Terms
}
\description{
\describe{
\item{Database}{Scopus and WoS (Web of Science)  are abstract and citation database of peer-reviewed literature:
scientific journals, books and conference proceedings. Scopus: https://www.scopus.com/home.uri; WoS: https://www.webofscience.com/wos/alldb/basic-search}
\item{Topic}{topic is the high level focal concept that the search focusses on, we have practices (list in parenthesis), outcomes (list in parenthesis) and geographic region}
\item{Keywords}{keywords used in the peer-reviewed literature}
\item{Date}{the year of publication}
...
}
}
\keyword{datasets}
