% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERA_SoilGrids18}
\alias{ERA_SoilGrids18}
\title{SoilGrids Parameters}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1520 rows and 168 columns.
}
\source{
https://www.isda-africa.com/isdasoil/?location=13.812\%2C-21.671\%2C43.677\%2C-0.051 & https://github.com/ISRICWorldSoil/SoilGrids250m/blob/master/grids/models/TAXNWRB/TAXNWRB_legend.csv & https://github.com/ISRICWorldSoil/SoilGrids250m/blob/master/grids/models/TAXOUSDA/TAXOUSDA_legend.csv
}
\usage{
ERA_SoilGrids18
}
\description{
\describe{
Using ISRIC World Soil Information, soil data were downloaded and summarized for each unique ERA locations and its buffer of spatial
uncertainty.
\item{Latitude}{latitude of study location in decimal degrees}
\item{Longitude}{longitude of study location in decimal degrees}
\item{Buffer}{estimation in m of the spatial uncertainty for a point location}
\item{Site.Key}{longitude, latitude and buffer for the site}
\item{ACDWRB_M_ss_250m.Mean}{mean values of Acid sub-soils grade. Grade of a sub-soil being acid e.g. having a pH < 5 and low BS}
\item{ACDWRB_M_ss_250m.SD}{standard deviation of Acid sub-soils grade. Grade of a sub-soil being acid e.g. having a pH < 5 and low BS}
\item{ACDWRB_M_ss_250m.Quantiles}{quantiles of Acid sub-soils grade. Grade of a sub-soil being acid e.g. having a pH < 5 and low BS}
\item{AWCh1_M_sl1_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a dept of 0-5cm. Unit = percent}
\item{AWCh1_M_sl1_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a dept of 0-5cm. Unit = percent}
\item{AWCh1_M_sl1_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a dept of 0-5cm. Unit = percent}
\item{AWCh1_M_sl2_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 5-15cm. Unit = percent}
\item{AWCh1_M_sl2_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 5-15cm. Unit = percent}
\item{AWCh1_M_sl2_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 5-15cm. Unit = percent}
\item{AWCh1_M_sl3_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 15-30cm. Unit = percent}
\item{AWCh1_M_sl3_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 15-30cm. Unit = percent}
\item{AWCh1_M_sl3_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h1 (moisture potential in kPa, e.g. -10 (pF 2.0)) at a depth of 15-30cm. Unit = percent}
\item{AWCh2_M_sl1_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 0-5cm. Unit = percent}
\item{AWCh2_M_sl1_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 0-5cm. Unit = percent}
\item{AWCh2_M_sl1_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 0-5cm. Unit = percent}
\item{AWCh2_M_sl2_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 5-15cm. Unit = percent}
\item{AWCh2_M_sl2_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 5-15cm. Unit = percent}
\item{AWCh2_M_sl2_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 5-15cm. Unit = percent}
\item{AWCh2_M_sl3_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 15-30cm. Unit = percent}
\item{AWCh2_M_sl3_250m.SD}{standard deviation of}{Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 15-30cm. Unit = percent}
\item{AWCh2_M_sl3_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h2 (moisture potential in kPa, e.g. -20 (pF 2.3)) at a depth of 15-30cm. Unit = percent}
\item{AWCh3_M_sl1_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 0-5cm. Unit = percent}
\item{AWCh3_M_sl1_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 0-5cm. Unit = percent}
\item{AWCh3_M_sl1_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 0-5cm. Unit = percent}
\item{AWCh3_M_sl2_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 5-15cm. Unit = percent}
\item{AWCh3_M_sl2_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 5-15cm. Unit = percent}
\item{AWCh3_M_sl2_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 5-15cm. Unit = percent}
\item{AWCh3_M_sl3_250m.Mean}{mean value of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 15-30cm. Unit = percent}
\item{AWCh3_M_sl3_250m.SD}{standard deviation of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 15-30cm. Unit = percent}
\item{AWCh3_M_sl3_250m.Quantiles}{quantiles of Available soil water capacity (volumetric fraction) for h3 (moisture potential in kPa, e.g. -31.6  (pF 2.5)) at a dept of 15-30cm. Unit = percent}
\item{AWCtS_M_sl1_250m.Mean}{mean value of Saturated water content (volumetric fraction) for tS at a depth of 0-5cm. Unit = percent}
\item{AWCtS_M_sl1_250m.SD}{standard deviation of Saturated water content (volumetric fraction) for tS at a depth of 0-5cm. Unit = percent}
\item{AWCtS_M_sl1_250m.Quantiles}{quantiles of Saturated water content (volumetric fraction) for tS at a depth of 0-5cm. Unit = percent}
\item{AWCtS_M_sl2_250m.Mean}{mean value of Saturated water content (volumetric fraction) for tS at a depth of 5-15cm. Unit = percent}
\item{AWCtS_M_sl2_250m.SD}{standard deviation of Saturated water content (volumetric fraction) for tS at a depth of 5-15cm. Unit = percent}
\item{AWCtS_M_sl2_250m.Quantiles}{quantiles of Saturated water content (volumetric fraction) for tS at a depth of 5-15cm. Unit = percent}
\item{AWCtS_M_sl3_250m.Mean}{mean value of Saturated water content (volumetric fraction) for tS at a depth of 15-30cm. Unit = percent}
\item{AWCtS_M_sl3_250m.SD}{standard deviation of Saturated water content (volumetric fraction) for tS at a depth of 15-30cm. Unit = percent}
\item{AWCtS_M_sl3_250m.Quantiles}{quantiles of Saturated water content (volumetric fraction) for tS at a depth of 15-30cm. Unit = percent}
\item{BDRICM_M_250m.Mean}{mean value of Depth to bedrock (R horizon) up to 200 cm. Unit = cm}
\item{BDRICM_M_250m.SD}{standard deviation of Depth to bedrock (R horizon) up to 200 cm. Unit = cm}
\item{BDRICM_M_250m.Quantiles}{quantiles of Depth to bedrock (R horizon) up to 200 cm. Unit = cm}
\item{BDRLOG_M_250m.Mean}{mean value of the Probability of occurrence of R horizon. Unit = percent}
\item{BDRLOG_M_250m.SD}{standard deviation of the  Probability of occurrence of R horizon. Unit = percent}
\item{BDRLOG_M_250m.Quantiles}{quantile of the Probability of occurrence of R horizon. Unit = percent}
\item{BDTICM_M_250m.Mean}{mean value of the Absolute depth to bedrock. Unit = cm}
\item{BDTICM_M_250m.SD}{standard deviation of the Absolute depth to bedrock. Unit = cm}
\item{BDTICM_M_250m.Quantiles}{quantile of the Absolute depth to bedrock. Unit = cm}
\item{BLDFIE_M_sl1_250m.Mean}{mean value of Bulk density (fine earth) in kg per cubic meter at a depth of 0-5cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl1_250m.SD}{standard deviation of Bulk density (fine earth) in kg per cubic meter at a depth of 0-5cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl1_250m.Quantiles}{quantiles of Bulk density (fine earth) in kg per cubic meter at a depth of 0-5cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl2_250m.Mean}{mean value of Bulk density (fine earth) in kg per cubic meter at a depth of 5-15cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl2_250m.SD}{standard deviation of Bulk density (fine earth) in kg per cubic meter at a depth of 5-15cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl2_250m.Quantiles}{quantiles of Bulk density (fine earth) in kg per cubic meter at a depth of 5-15cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl3_250m.Mean}{mean value of Bulk density (fine earth) in kg per cubic meter at a depth of 15-30cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl3_250m.SD}{standard deviation of Bulk density (fine earth) in kg per cubic meter at a depth of 15-30cm. Unit = kg/cubic-m}
\item{BLDFIE_M_sl3_250m.Quantiles}{quantiles of Bulk density (fine earth) in kg per cubic meter at a depth of 15-30cm. Unit = kg/cubic-m}
\item{CECSOL_M_sl1_250m.Mean}{mean of Cation Exchange Capacity of soil at a depth of 0-5cm. Unit = cmol/kg}
\item{CECSOL_M_sl1_250m.SD}{standard deviation of Cation Exchange Capacity of soil at a depth of 0-5cm. Unit = cmol/kg}
\item{CECSOL_M_sl1_250m.Quantiles}{quantiles of Cation Exchange Capacity of soil at a depth of 0-5cm. Unit = cmol/kg}
\item{CECSOL_M_sl2_250m.Mean}{mean of Cation Exchange Capacity of soil at a depth of 5-15cm. Unit = cmol/kg}
\item{CECSOL_M_sl2_250m.SD}{standard deviation of Cation Exchange Capacity of soil at a depth of 5-15cm. Unit = cmol/kg}
\item{CECSOL_M_sl2_250m.Quantiles}{Cation Exchange Capacity of soil at a depth of 5-15cm. Unit = cmol/kg}
\item{CECSOL_M_sl3_250m.Mean}{mean of Cation Exchange Capacity of soil at a depth of 15-30cm. Unit = cmol/kg}
\item{CECSOL_M_sl3_250m.SD}{standard deviation of Cation Exchange Capacity of soil at a depth of 15-30cm. Unit = cmol/kg}
\item{CECSOL_M_sl3_250m.Quantiles}{standard deviation of Cation Exchange Capacity of soil at a depth of 15-30cm. Unit = cmol/kg}
\item{CLYPPT_M_sl1_250m.Mean}{mean of Weight percentage of the clay particles (<0.0002 mm) at a depth of 0-5cm. Unit = percent}
\item{CLYPPT_M_sl1_250m.SD}{standard deviation of Weight percentage of the clay particles (<0.0002 mm) at a depth of 0-5cm. Unit = percent}
\item{CLYPPT_M_sl1_250m.Quantiles}{quantiles of Weight percentage of the clay particles (<0.0002 mm) at a depth of 0-5cm. Unit = percent}
\item{CLYPPT_M_sl2_250m.Mean}{mean of Weight percentage of the clay particles (<0.0002 mm) at a depth of 5-15cm. Unit = percent}
\item{CLYPPT_M_sl2_250m.SD}{standard deviation of Weight percentage of the clay particles (<0.0002 mm) at a depth of 5-15cm. Unit = percent}
\item{CLYPPT_M_sl2_250m.Quantiles}{quantiles of Weight percentage of the clay particles (<0.0002 mm) at a depth of 5-15cm. Unit = percent}
\item{CLYPPT_M_sl3_250m.Mean}{mean of Weight percentage of the clay particles (<0.0002 mm) at a depth of 15-30cm. Unit = percent}
\item{CLYPPT_M_sl3_250m.SD}{standard deviation of Weight percentage of the clay particles (<0.0002 mm) at a depth of 15-30cm. Unit = percent}
\item{CLYPPT_M_sl3_250m.Quantiles}{quantiles of Weight percentage of the clay particles (<0.0002 mm) at a depth of 15-30cm. Unit = percent}
\item{CRFVOL_M_sl1_250m.Mean}{mean value of Volumetric percentage of coarse fragments (>2 mm) at a depth of 0-5cm. Unit = percent}
\item{CRFVOL_M_sl1_250m.SD}{standard deviation of Volumetric percentage of coarse fragments (>2 mm) at a depth of 0-5cm. Unit = percent}
\item{CRFVOL_M_sl1_250m.Quantiles}{quantiles of Volumetric percentage of coarse fragments (>2 mm) at a depth of 0-5cm. Unit = percent}
\item{CRFVOL_M_sl2_250m.Mean}{mean value of Volumetric percentage of coarse fragments (>2 mm) at a depth of 5-15cm. Unit = percent}
\item{CRFVOL_M_sl2_250m.SD}{standard deviation of Volumetric percentage of coarse fragments (>2 mm) at a depth of 5-15cm. Unit = percent}
\item{CRFVOL_M_sl2_250m.Quantiles}{quantiles of Volumetric percentage of coarse fragments (>2 mm) at a depth of 5-15cm. Unit = percent}
\item{CRFVOL_M_sl3_250m.Mean}{mean value of Volumetric percentage of coarse fragments (>2 mm) at a depth of 15-30cm. Unit = percent}
\item{CRFVOL_M_sl3_250m.SD}{standard deviation of Volumetric percentage of coarse fragments (>2 mm) at a depth of 15-30cm. Unit = percent}
\item{CRFVOL_M_sl3_250m.Quantiles}{quantiles of Volumetric percentage of coarse fragments (>2 mm) at a depth of 15-30cm. Unit = percent}
\item{gyga_af_erzd_limfactor__m_1km.Mean}{mean of rootable depth limiting soil factor for maize, mapped at 1km resolution}
\item{gyga_af_erzd_limfactor__m_1km.SD}{standard deviation of rootable depth limiting soil factor for maize, mapped at 1km resolution}
\item{gyga_af_erzd_limfactor__m_1km.Quantiles}{quantiles of rootable depth limiting soil factor for maize, mapped at 1km resolution}
\item{OCSTHA_M_sd1_250m.Mean}{mean value of Soil organic carbon stock in tons per ha at a depth of 0-5cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd1_250m.SD}{standard deviation of Soil organic carbon stock in tons per ha at a depth of 0-5cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd1_250m.Quantiles}{quantiles of Soil organic carbon stock in tons per ha at a depth of 0-5cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd2_250m.Mean}{mean value of Soil organic carbon stock in tons per ha at a depth of 5-15cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd2_250m.SD}{standard deviation of Soil organic carbon stock in tons per ha at a depth of 5-15cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd2_250m.Quantiles}{quantiles of Soil organic carbon stock in tons per ha at a depth of 5-15cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd3_250m.Mean}{mean value of Soil organic carbon stock in tons per ha at a depth of 15-30cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd3_250m.SD}{standard deviation of Soil organic carbon stock in tons per ha at a depth of 15-30cm. Unit = tonnes/ha}
\item{OCSTHA_M_sd3_250m.Quantiles}{quantiles of Soil organic carbon stock in tons per ha at a depth of 15-30cm. Unit = tonnes/ha}
\item{ORCDRC_M_sl1_250m.Mean}{mean value of soil organic carbon content in permille at a depth of 0-5cm. Unit = g/kg}
\item{ORCDRC_M_sl1_250m.SD}{standard deviation of soil organic carbon content in permille at a depth of 0-5cm. Unit = g/kg}
\item{ORCDRC_M_sl1_250m.Quantiles}{quantiles of soil organic carbon content in permille at a depth of 0-5cm. Unit = g/kg}
\item{ORCDRC_M_sl2_250m.Mean}{mean value of soil organic carbon content in permille at a depth of 5-15cm. Unit = g/kg}
\item{ORCDRC_M_sl2_250m.SD}{standard deviation of soil organic carbon content in permille at a depth of 5-15cm. Unit = g/kg}
\item{ORCDRC_M_sl2_250m.Quantiles}{quantiles of soil organic carbon content in permille at a depth of 5-15cm. Unit = g/kg}
\item{ORCDRC_M_sl3_250m.Mean}{mean value of soil organic carbon content in permille at a depth of 15-30cm. Unit = g/kg}
\item{ORCDRC_M_sl3_250m.SD}{standard deviation of soil organic carbon content in permille at a depth of 15-30cm. Unit = g/kg}
\item{ORCDRC_M_sl3_250m.Quantiles}{quantiles of soil organic carbon content in permille at a depth of 15-30cm. Unit = g/kg}
\item{PHIHOX_M_sl1_250m.Mean}{mean value of pH index measured in water solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIHOX_M_sl1_250m.SD}{standard deviation of pH index measured in water solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIHOX_M_sl1_250m.Quantiles}{quantiles of pH index measured in water solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIHOX_M_sl2_250m.Mean}{mean value of pH index measured in water solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIHOX_M_sl2_250m.SD}{pH index measured in water solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIHOX_M_sl2_250m.Quantiles}{pH index measured in water solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIHOX_M_sl3_250m.Mean}{mean value of pH index measured in water solution at a depth of 15-30cm. Unit = pH x10}
\item{PHIHOX_M_sl3_250m.SD}{pH index measured in water solution at a depth of 15-30cm. Unit = pH x10}
\item{PHIHOX_M_sl3_250m.Quantiles}{pH index measured in water solution at a depth of 15-30cm. Unit = pH x10}
\item{PHIKCL_M_sl1_250m.Mean}{mean value of pH index measured in KCl solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIKCL_M_sl1_250m.SD}{standard deviation of pH index measured in KCl solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIKCL_M_sl1_250m.Quantiles}{quantiles of pH index measured in KCl solution at a depth of 0-5cm. Unit = pH x10}
\item{PHIKCL_M_sl2_250m.Mean}{mean value of pH index measured in KCl solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIKCL_M_sl2_250m.SD}{standard deviation of pH index measured in KCl solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIKCL_M_sl2_250m.Quantiles}{quantiles of pH index measured in KCl solution at a depth of 5-15cm. Unit = pH x10}
\item{PHIKCL_M_sl3_250m.Mean}{mean value of pH index measured in KCl solution at a depth of 15-30cm. Unit = pH x10}
\item{PHIKCL_M_sl3_250m.SD}{standard deviation of pH index measured in KCl solution at a depth of 15-30cm. Unit = pH x10}
\item{PHIKCL_M_sl3_250m.Quantiles}{quantiles of pH index measured in KCl solution at a depth of 15-30cm. Unit = pH x10}
\item{SLGWRB_250m.Mean}{mean value of Sodic soil grade based on WRB soil types and soil pH}
\item{SLGWRB_250m.SD}{standard deviation of Sodic soil grade based on WRB soil types and soil pH}
\item{SLGWRB_250m.Quantiles}{quantiles of Sodic soil grade based on WRB soil types and soil pH}
\item{SLTPPT_M_sl1_250m.Mean}{mean value of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 0-5cm. Unit = percent}
\item{SLTPPT_M_sl1_250m.SD}{standard deviation of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 0-5cm. Unit = percent}
\item{SLTPPT_M_sl1_250m.Quantiles}{quantiles of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 0-5cm. Unit = percent}
\item{SLTPPT_M_sl2_250m.Mean}{mean value of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 5-15cm. Unit = percent}
\item{SLTPPT_M_sl2_250m.SD}{standard deviation of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 5-15cm. Unit = percent}
\item{SLTPPT_M_sl2_250m.Quantiles}{quantiles of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 5-15cm. Unit = percent}
\item{SLTPPT_M_sl3_250m.Mean}{mean value of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 15-30cm. Unit = percent}
\item{SLTPPT_M_sl3_250m.SD}{standard deviation of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 15-30cm. Unit = percent}
\item{SLTPPT_M_sl3_250m.Quantiles}{quantiles of Weight percentage of the silt particles (0.0002-0.05 mm) at a depth of 0-5cm. Unit = percent}
\item{SNDPPT_M_sl1_250m.Mean}{mean value of Weight percentage of the sand particles (0.05-2 mm) at a depth of 0-5cm. Unit = percent}
\item{SNDPPT_M_sl1_250m.SD}{standard deviation of Weight percentage of the sand particles (0.05-2 mm) at a depth of 0-5cm. Unit = percent}
\item{SNDPPT_M_sl1_250m.Quantiles}{quantiles of Weight percentage of the sand particles (0.05-2 mm) at a depth of 0-5cm. Unit = percent}
\item{SNDPPT_M_sl2_250m.Mean}{mean value of Weight percentage of the sand particles (0.05-2 mm) at a depth of 5-15cm. Unit = percen}
\item{SNDPPT_M_sl2_250m.SD}{standard deviation of Weight percentage of the sand particles (0.05-2 mm) at a depth of 5-15cm. Unit = percent}
\item{SNDPPT_M_sl2_250m.Quantiles}{quantiles of Weight percentage of the sand particles (0.05-2 mm) at a depth of 5-15cm. Unit = percent}
\item{SNDPPT_M_sl3_250m.Mean}{mean value of Weight percentage of the sand particles (0.05-2 mm) at a depth of 15-30cm. Unit = percen}
\item{SNDPPT_M_sl3_250m.SD}{standard deviation of Weight percentage of the sand particles (0.05-2 mm) at a depth of 15-30cm. Unit = percent}
\item{SNDPPT_M_sl3_250m.Quantiles}{quantiles of Weight percentage of the sand particles (0.05-2 mm) at a depth of 15-30cm. Unit = percent}
\item{TAXNWRB_250m.Mode}{mode of Predicted WRB 2006 subgroup classes}
\item{TAXOUSDA_250m.Mode}{predicted WRB 2006 subgroup classes}
\item{TEXMHT_M_sl1_250m.Mean}{mean value of Texture class (USDA system) at a depth of 0-5cm. Unit = factor}
\item{TEXMHT_M_sl1_250m.SD}{standard deviation of Texture class (USDA system) at a depth of 0-5cm. Unit = factor}
\item{TEXMHT_M_sl1_250m.Quantiles}{quantiles of Texture class (USDA system) at a depth of 0-5cm. Unit = factor}
\item{TEXMHT_M_sl2_250m.Mean}{mean value of Texture class (USDA system) at a depth of 5-15cm. Unit = factor}
\item{TEXMHT_M_sl2_250m.SD}{standard deviation of Texture class (USDA system) at a depth of 5-15cm. Unit = factor}
\item{TEXMHT_M_sl2_250m.Quantiles}{quantiles of Texture class (USDA system) at a depth of 5-15cm. Unit = factor}
\item{TEXMHT_M_sl3_250m.Mean}{mean value of Texture class (USDA system) at a depth of 15-30cm. Unit = factor}
\item{TEXMHT_M_sl3_250m.SD}{standard deviation of Texture class (USDA system) at a depth of 15-30cm. Unit = factor}
\item{TEXMHT_M_sl3_250m.Quantiles}{quantiles of Texture class (USDA system) at a depth of 15-30cm. Unit = factor}
\item{WWP_M_sl1_250m.Mean}{mean value of available soil water capacity (volumetric fraction) until wilting point at a depth of 0-5cm. Unit = percent}
\item{WWP_M_sl1_250m.SD}{standard deviation of available soil water capacity (volumetric fraction) until wilting point at a depth of 0-5cm. Unit = percent}
\item{WWP_M_sl1_250m.Quantiles}{quantiles of available soil water capacity (volumetric fraction) until wilting point at a depth of 0-5cm. Unit = percent}
\item{WWP_M_sl2_250m.Mean}{mean value of available soil water capacity (volumetric fraction) until wilting point at a depth of 5-15cm. Unit = percent}
\item{WWP_M_sl2_250m.SD}{standard deviation of available soil water capacity (volumetric fraction) until wilting point at a depth of 5-15cm. Unit = percent}
\item{WWP_M_sl2_250m.Quantiles}{quantiles of available soil water capacity (volumetric fraction) until wilting point at a depth of 5-15cm. Unit = percent}
\item{WWP_M_sl3_250m.Mean}{mean value of available soil water capacity (volumetric fraction) until wilting point at a depth of 15-30cm. Unit = percent}
\item{WWP_M_sl3_250m.SD}{standard deviation of available soil water capacity (volumetric fraction) until wilting point at a depth of 15-30cm. Unit = percent}
\item{WWP_M_sl3_250m.Quantiles}{quantiles of available soil water capacity (volumetric fraction) until wilting point at a depth of 15-30cm. Unit = percent}
...
}
}
\keyword{datasets}
