% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAAlphaPlot.R
\name{ERAAlphaPlot}
\alias{ERAAlphaPlot}
\title{Plot ERA Data Locations with Alpha Density Layer}
\usage{
ERAAlphaPlot(
  Data = ERA.Compiled,
  Background = NA,
  Background.Labs = NA,
  Background.Cols = NA,
  Background.Title = NA,
  alpha.bandwidth = 4,
  Showpoints = T,
  Low = "black",
  Mid = "grey30",
  High = "white",
  Point.Col = "Black",
  ALevel = NA
)
}
\arguments{
\item{Data}{A data.frame or data.table with decimal degree point locations in columns named \code{Latitude} & \code{Longitude}. NA values are not permitted.}

\item{Background}{A raster file that forms the background of the map. This file is cropped to the extent of Africa. A default farming system map is supplied from the
\href{https://doi.org/10.7910/DVN/G4TBLF/7RIFT5}{CELL5M} dataset.}

\item{Background.Labs}{A vector of names that correspond to the background raster values (in numerical order of the raster values).}

\item{Background.Cols}{A vector of colours that correspond to the background raster values (in numerical order of the raster values).}

\item{Background.Title}{The name of the background raster.}

\item{alpha.bandwidth}{A multiplicative bandwidth adjustment for use with the \link[ggplot2]{geom_density_2d} function. Higher values increase the size of the transparent area around
high data density regions.}

\item{Showpoints}{Logical \code{T/F}. If \code{TRUE} points locations are added to the plot.}

\item{Low}{Colour of low data density areas. Default = \code{black}.}

\item{Mid}{Colour of medium data density areas. Default = \code{grey30}.}

\item{High}{Colour of high data density areas. Default = \code{white}.}

\item{Point.Col}{Colour of points. Default = \code{black}.}

\item{ALevel}{If \code{NA} then the dataset is subset to unique values of \code{Latitude} and \code{Longitude}. If a vector of column names is supplied then then the dataset
is subset to unique values of \code{Latitude} and \code{Longitude} plus these columns. Default = \code{NA}.}
}
\value{
AlphaD.Plot returns a \code{ggplot} object showing a map of point data density.
}
\description{
This function plots the locations of ERA observations on a map of Africa, with a supplied background raster. A density raster is overlain which is transparent in high density areas (many ERA observations) and opaque in low density areas (few ERA observations). The bandwidth of alpha in the density raster can be adjusted using the \code{alpha.bandwidth}
parameter.
}
\details{
The map of Africa is supplied using the  \link[rworldmap]{getMap} function.
}
