% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAChordPlot.R
\name{ERAChordPlot}
\alias{ERAChordPlot}
\title{Create Chord Plots}
\usage{
ERAChordPlot(
  Chord.Data,
  Type = "Chord",
  Alpha = 0.5,
  Value.Var = "yi",
  Value.Mid = 0,
  Cont.Pallete = NA,
  Legend.Tit = "",
  Rotate.Labs = F,
  Lab.Adjust = 0,
  Lab.Cex = 0.2,
  Axis.Cex = 0.3,
  Reduce = 0.01,
  Cat.Pal = NA
)
}
\arguments{
\item{Chord.Data}{A \code{data.table} object output by the \code{ERAg::PrepareChordData}function}

\item{Type}{A character vector of either \code{"Chord"},\code{"Grid"} or \code{"Category"}. If \code{Type="Grid"} chords are coloured by the \code{from} objects in the plot, if \code{Type="Chord"}
chords are colored by continuous numeric field specified by \code{Value.Var}, and if \code{Type="Category"}chords are coloured by a categorical variable.}

\item{Alpha}{A numeric value between 0 and 1 indicating the transparency of chords}

\item{Value.Var}{A character vector of either \code{"yi"},\code{"vote"} or \code{"cat"}. If \code{Type="Chord"} then this argument chooses the value statistic that colours chords, \code{"yi"}
= response ratio and \code{"vote"} = vote count. If \code{Type="Category"} then this argument should be set to \code{"cat"}; a column of class \code{character} in \code{Chord.Data} containing a categorical variable.}

\item{Value.Mid}{A numeric vector of length one; choose the mid-value of the colour palette, for response ratios (\code{Value.Var=="yi"}) this is typically 0}

\item{Cont.Pallete}{A character vector of length two describing two \code{brewer.pal}palettes from\verb{rownames(RColorBrewer::brewer.pal.info)[RColorBrewer::brewer.pal.info$category=="seq"])}.
Set to \code{NA} to use default palettes}

\item{Legend.Tit}{A character vector of length one; title of the legend}

\item{Rotate.Labs}{Logical \code{T/F}. If \code{T} labels are rotated perpendicular to the plot track}

\item{Lab.Adjust}{A numeric vector of length one; adjust the spacing between rotated labels and the track}

\item{Lab.Cex}{A numeric vector of length one; adjusts the size of sector labels}

\item{Axis.Cex}{A numeric vector of length one; adjusts the size of track axis labels}

\item{Reduce}{A numeric vector of length one; a proportion between 0-1, chords widths below this proportion are excluded from the plot}

\item{Cat.Pal}{A character vector of length = \code{length(unique(Chord.Data$cat))}; the vector should contain colour values corresponding to the values in
\code{Chord.Data$cat} which should form the vector names, for example  \code{Cat.Pal=setNames(c("blue", "green","Red"), c("A", "B","C"))}.}
}
\value{
Returns a \link[circlize]{chordDiagram} plot capture using \code{recordPlot()}.
}
\description{
The \code{ERAChordPlot} function plots the output of the \code{ERAg::PrepareChordData} function to visualise the relationship between ERA objects.
}
\details{
This function uses the \link[circlize]{chordDiagram} to create the chord plot and also requires the \code{ComplexHeatmap} package to generate legends.
The latter can be installed using \verb{devtools::install_github("jokergoo/ComplexHeatmap").}
}
