% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAHexPlot.R
\name{ERAHexPlot}
\alias{ERAHexPlot}
\title{Plot ERA Data Locations with Hex Density Layer}
\usage{
ERAHexPlot(
  Data,
  Low = "grey10",
  Mid = "grey50",
  High = "black",
  Point.Col = "yellow",
  Do.Log = "Yes",
  Showpoints = "No",
  ALevel = NA
)
}
\arguments{
\item{Data}{A data.frame or data.table with decimal degree point locations in columns named \code{Latitude} & \code{Longitude}. NA values are not permitted.}

\item{Low}{Colour of low data density areas. Default = \code{grey10}.}

\item{Mid}{Colour of medium data density areas. Default = \code{grey50}.}

\item{High}{Colour of high data density areas. Default = \code{black}.}

\item{Point.Col}{Colour of points. Default = \code{yellow}.}

\item{Do.Log}{A character vector of \code{Yes}  or \code{No}. If \code{Yes} the shading of hexagons uses logged counts as \code{stat(log(count))}. Default = \code{Yes}.}

\item{Showpoints}{A character vector of \code{Yes}  or \code{No}. If \code{Yes} point locations are plotted on the map. Default = \code{No}.}

\item{ALevel}{If \code{NA} then the dataset is subset to unique values of \code{Latitude} and \code{Longitude}. If a vector of column names is supplied then the dataset is subset to
unique values of \code{Latitude} and \code{Longitude} plus these columns. Default = \code{NA}.}
}
\value{
ERAHexPlot returns a \code{ggplot} object showing a map of point density
}
\description{
This function plots the density of ERA observations over a map of Africa. A hexagonal density layer is overlain using the \link[ggplot2]{geom_hex} function and points can be displayed.
}
\details{
The map of Africa is supplied using the  \link[rworldmap]{getMap} function.
}
