% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAMatrixPlot.R
\name{ERAMatrixPlot}
\alias{ERAMatrixPlot}
\title{Data Availability Matrix Plot}
\usage{
ERAMatrixPlot(
  Data,
  DoLog = "No",
  Palette = "plasma",
  FlipXY = "No",
  RotYlab = "Yes",
  ALevel = "Observation"
)
}
\arguments{
\item{Data}{a \code{list} returned from the from the \code{ERA::PrepareXY} function}

\item{DoLog}{A character vector of length one; accepted values: \code{Yes}  or \code{No}; default = \code{No}; if \code{Yes} fill values are natural log transformed.}

\item{Palette}{A character vector of length one; accepted values: \code{magma}, \code{inferno}, \code{plasma} or \code{viridis}; default = \code{plasma}; sets the palette of
of the matrix fill using the \link[ggplot]{scale_fill_viridis} function}

\item{FlipXY}{A character vector of length one; accepted values: \code{Yes}  or \code{No}; default = \code{No}; if \code{Yes} the x and y axes of the plot are swapped}

\item{RotYlab}{A character vector of length one; accepted values: \code{Yes}  or \code{No}; default = \code{Yes}; if \code{Yes} the y axis labels of the plot are rotated 90 degrees}

\item{ALevel}{A character vector of length one that labels the legend of the plot, this should refer to any data aggregation by study or location that has occurred (or not); default = \code{Observation}}
}
\value{
\code{ERAMatrixPlot} returns a  \link[ggplot]{geom_tile} plot
}
\description{
Plots a matrix shaded for data availability using the \link[ggplot]{geom_tile} function for an object returned from the \code{PrepareXY} function.
}
\details{
The fill of and numbers in the matrix tiles are the number of rows for each combination of values in \code{Axis1} and \code{Axis2} in the \code{Data} element of the
list returned by \code{ERA::PrepareXY}.

Note \code{PrepareXY} selects of two of the three higher order ERA concepts of  \strong{practice}, \strong{outcome} and \strong{product}, counts for the selected concepts
include all levels of the non-selected concept. To explore results within in a non-selected concept you will need to subset the \code{Data} element of the
list returned by \code{PrepareXY} on the \code{Group} column.
}
