% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAWeightR.R
\name{ERAWeight}
\alias{ERAWeight}
\title{Calculate Weightings}
\usage{
ERAWeight(Data, Grouping.Cols)
}
\arguments{
\item{Data}{An ERA data.table containing columns \code{Reps} and \code{Code}.}

\item{Grouping.Cols}{A vector of columns names; these are grouping variables and weightings are calculated for each combination of their values. Default = \code{NA}.}
}
\value{
ERAWeights returns the input \code{data.table} with two appended columns:
\itemize{
\item \code{N.Obs.Study} = the number of observations contributed by a study (i.e., \code{Code}) within combinations of the grouping variables specified.
\item \code{Weight.Study} = a numeric weighting.
}
}
\description{
This simple function will weight ERA observations based on replicates (\code{Reps} column) and the number of observations contributed by a study (\code{N.Obs.Study})
within combinations of the grouping variables specified.
}
\details{
For each observations weightings are calculated using the formula \code{(Rep^2/(2*Rep))/N.Obs.Study} where \code{Rep} is the number of experimental replicates
and \code{N.Obs.Study} is the number of observations contributed from the observation's study for the grouping variables provided.
}
