% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAWide2Long.R
\name{ERAWide2Long}
\alias{ERAWide2Long}
\title{Convert ERA from wide to long format}
\usage{
ERAWide2Long(Data)
}
\arguments{
\item{Data}{The ERA.Compiled dataset or a subset of it.}
}
\value{
A data.table where control vs experimental treatment comparisons have been removed. Each unique outcome observation (control or experimental) is a row in this table.
}
\description{
This function takes wide format ERA dataset where a control is compared to an experimental treatment and converts it to a long format where each row is a single unique outcome observation.
}
