% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPDayData.R
\name{EstPDayData}
\alias{EstPDayData}
\title{Estimate missing planting dates from existing values in ERA}
\usage{
EstPDayData(DATA)
}
\arguments{
\item{DATA}{An ERA dataset (e.g. \code{ERA.Compiled})}
}
\value{
The following fields are appended to the input dataset:
\code{Data.PS.Date} = planting start date estimated from the dataset provided (Date)
\code{Data.PE.Date} = planting end date estimated from the dataset provided (Date)
\code{Data.Date.Acc} = a concantenation of the number of decimals places used in latitude/longitude matching and whether matching was conducted using
\code{Product} (\code{P.Sub}) or \code{Product.Subtype} (\code{P}), for example \code{3-P.Sub}.
}
\description{
For NA values of planting date in the dataset supplied the function attempts to substitute values for the same \code{Product}/\code{Product.Subtype} and growing season
which are spatially nearby.
}
\details{
#' The function searches iteratively within five levels of increasing distance corresponding to latitude and longitude recorded from 5 to 1 decimal
places. If there is more than one planting date value for a \code{Product} x season (\code{M.Year}) combination then values are averaged. If no corresponding
planting dates are found matching on product then matching is attempted on \code{Product.Subtype} (e.g. cereals or legumes).
}
