% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPDayRain.R
\name{EstPDayRain}
\alias{EstPDayRain}
\title{Estimate uncertain planting dates from rainfall data}
\usage{
EstPDayRain(
  Data,
  ID,
  Rain.Data,
  Rain.Data.Name,
  Rain.Field,
  DaysBefore = 0,
  MultiplyWin = 1,
  Window = c(14, 14),
  Widths = c(2, 3, 3),
  Rain.Thresholds,
  Uncertainty.Min = 7,
  Uncertainty.Max = 90,
  Add.Values = T,
  Use.Data.Dates = F
)
}
\arguments{
\item{Data}{An ERA dataset (e.g. \code{ERA.Compiled})}

\item{ID}{A character vector of length one containing the column name for a unique id field naming each location in the dataset provided.}

\item{Rain.Data}{Data.frame or data.table of daily rainfall with a field of class \code{Date} named \code{Date}, a numeric field named as per the \code{Rain.Field} argu containing daily rainfall amount in mm and ID field matching Data object supplied.}

\item{Rain.Data.Name}{A character vector of length one containing the name of the rainfall dataset (e.g. "CHIRPS")}

\item{Rain.Field}{A character vector of length one containing the name of the rainfall field in the \code{Rain.Data} object supplied (e.g. \code{Rain})}

\item{DaysBefore}{An integer vector of length one; when searching for rainfall events in-between the planting start/end dates specified, extend the planting start date backwards by a number of days}

\item{MultiplyWin}{An numeric vector of length one; this parameter is a proportion that changes the size of the difference between plant start and plant end, 1= no change}

\item{Window}{An integer vector of max length two which specifies in days up to two addition temporal periods beyond the initial temporal such period. Can be set to NA if no additional temporal search periods are required.}

\item{Widths}{An integer vector of length equivalent to 1 + length(\code{Window}), max length = 3; the width of the scanning window in days within which rainfall is summed for the initial window of planting uncertainty plus each of the additional periods specified in the \code{Windows} argument.}

\item{Rain.Thresholds}{An integer vector of length equivalent to 1 + length(\code{Window}), max length = 3; the amount of rainfall that has to fall in the temporal windows considered.}

\item{Uncertainty.Min}{An integer vector of length one; the minimum acceptable difference between plant.start and plant.end in days}

\item{Uncertainty.Max}{An integer vector of length one; the maximum acceptable difference between plant.start and plant.end in days}

\item{Add.Values}{Logical \code{T/F}; if \code{T} append results to the Data object supplied and return, else return a separate table.}

\item{Use.Data.Dates}{Logical \code{T/F}; only relevant if planting date estimates have been added to the Data object using the \code{ERAg::Est.PDay.Data} function. If \code{T} then NA values in for Plant.Start and Plant.End are replaced by values estimated using the Est.PDay.Data function.}

\item{SaveDir}{A character vector of length one containing the path to the directory where the extracted data are to be saved. Set to NA if you do not want to save the returned dataset.}
}
\value{
The \code{EstPDayRain} function generates a field of class \code{Date} named \code{paste0("UnC.",Rain.Data.Name,".P.Date")} containing planting date estimates
for rows of \code{Data} where planting uncertainty is within the range specfied by the \code{Uncertainty.Min} and \code{Uncertainty.Max} arguments. If the \code{Add.Values}
arguments is \code{TRUE} then this field is appended to \code{Data} else it is returned in a separate \code{data.table.}
}
\description{
This function can be used to estimate planting datesfrom daily rainfall where there is uncertainty as calculated by substracting \code{Plant.Start} from \code{Plant.End}.
}
\details{
A daily rainfall dataset extracted for the data supplied in the \code{Data} argument needs to be supplied to this function, we recommend using
the \code{ExtractCHIRPS} function to generate rainfall data. In the \code{Rain.Data.Name} argument please supply the name of the rainfall dataset (e.g. \code{CHIRPS}).
Use the \code{Rain.Field} argument to specify the name of the column containing the rainfall amount. Matching between \code{Data} and \code{Rain.Data} uses the
\code{Daycount} and a location identity field as named using the \code{ID} argument, as such the datasets must share the same point of origin for calculation
of \code{Daycount} and the same name for the \code{ID} field.

Planting dates will estimated for rows in \code{Data} where the difference between \code{Plant.Start} and \code{Plant.End} is greater than or equal to the
\code{Uncertainty.Min} argument and less than or equal to the \code{Uncertainty.Max} argument.

\code{EstPDayRain} uses the  \link[zoo]{rollapply} function to sum rainfall within a scanning window, planting is assumed to occur the day \strong{after}
summed rainfall surpasses a threshold amount.

For each row of \code{Data} with appropriate planting uncertainty the function initially searches for rainfall events in \code{Rain.Data} in-between and including the corresponding \code{Plant.Start} and \code{Plant.End} dates.
This temporal search period can be modified using the \code{DaysBefore} and \code{MultiplyWin} arguments.\code{DaysBefore} extends the \code{Plant.Start} date backwards by a number of days.
\code{MultiplyWin} is a proportion that multiplies the difference between \code{Plant.Start} and \code{Plant.End} dates increasing the size of the period forwards in time.
If \code{Plant.Start = 2018-01-01} and \code{Plant.End = 2018-01-11} the difference between them is 10 days, if \code{MultiplyWin = 1.5} then \code{10*1.5=15} and the
end of the initial search window becomes \code{2018-01-01 + 15 = 2018-01-16}. The width (in days) of the \code{rollapply} scanning window is specified
by the \code{Width}argument and the amount of rainfall (mm) required to trigger planting within the scanning window is specified by the \code{Rain.Thresholds} argument.

Up to two additional temporal search periods after the above can be specified in days using the \code{Window} arguments, for each extra window added \code{Width} and
\code{Rain.Thresholds} arguments require values to added in sequence. If no trigger events are found in the initial scanning window then subsequent windows
are searched accordingly.
}
