% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstSLenData.R
\name{EstSLenData}
\alias{EstSLenData}
\title{Estimate missing season lengths from existing values in ERA}
\usage{
EstSLenData(DATA)
}
\arguments{
\item{DATA}{An ERA dataset (e.g. \code{ERA.Compiled})}
}
\value{
The following fields are appended to the input dataset:
\code{Data.SLen} = season length date estimated from the dataset provided (days)
\code{Data.SLen.Acc} = a concantenation of the number of decimals places used in latitude/longitude matching and whether matching was conducted using
\code{Product} (\code{P.Sub}) or \code{Product.Subtype} (\code{P}), for example \code{3-P.Sub}.
}
\description{
For NA values of season length in the ERA dataset supplied the function attempts to substitute values for the same \code{Product}/\code{Product.Subtype} and growing season
which are spatially nearby.
}
\details{
Season length should be present as a numeric field \code{SLen} in the supplied dataset. It can be calculated as the mean difference between mean harvest date (midpoint of \code{Harvest.Start} and \code{Harvest.End} Date columns) and mean planting date
(midpoint of \code{Plant.Start} and \code{Plant.End} Date columns).

The function searches iteratively within five levels of increasing distance corresponding to latitude and longitude recorded from 5 to 1 decimal
places. If there is more than one season length value for a \code{Product} x season (\code{M.Year}) combination then values are averaged. If no corresponding
season lengths are found matching on product then matching is attempted on \code{Product.Subtype} (e.g. cereals or legumes).
}
