% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUCodes}
\alias{EUCodes}
\title{EU Codes (Version 03.05.2021)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 355 rows and 42 columns.
}
\source{
ERA Project Team (2021)
}
\usage{
EUCodes
}
\description{
\describe{The organizational hierarchy of experimental units considered in ERA and their descriptions.
\item{EU}{unique alpha numeric code}
\item{Product.Type}{concepts in the highest level in the experimental unit hierarchy}
\item{Product.Type.Code}{currently a duplicate of Product.Type}
\item{Product.Subtype}{concepts in the intermediate level of the EU hierarchy. Product.Subtype is nested under Product.Type}
\item{Product.Subtype.Code}{short code corresponding to the Product.Subtype name for use in plotting}
\item{Product.Subtype.Code.Check}{currently all TRUE}
\item{Product}{concepts in the lowest and most detailed level of the outcome hierarchy.Product is nested under Product.Subtype}
\item{Product.Simple}{shorter and cleaner name for the Product}
\item{Product.Simple.Code}{short code corresponding to the Product.Simple name for use in plotting}
\item{Product.Code.Check}{currently all TRUE}
\item{Component}{the component of the Product that the experiment is collecting data on}
\item{Latin.Name}{the scientific name for the Product.Simple}
\item{Notes}{free text used to clarify any details of the Product and Product.Simple}
\item{Nfix}{used to identify whether the product is a nitrogen fixing plant. NA for animal products}
\item{Tree}{used to identify whether the product is a tree. NA for animal products}
\item{Mulched}{ERA mulch code based on the nitrogen fixing properties specified in Nfix. NA for animal products}
\item{Incorp}{ERA mulch incorporated code based on the nitrogen fixing properties specified in Nfix. NA for animal products}
\item{Unknown.Fate}{ERA code for unknown fate of the crop biomass. NA for animal products}
\item{ECOCROP.Name}{scientific name of the product exported from FAO's crop model- Ecocrop}
\item{ECOCROP.Notes}{free text to clarify andy details in ECOCROP.Name}
...
}
}
\keyword{datasets}
