% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractAgMERRA.R
\name{ExtractAgMERRA}
\alias{ExtractAgMERRA}
\title{Extract Climate Data from AgMERRA}
\usage{
ExtractAgMERRA(
  DATA,
  ID,
  AgMERRA_dir,
  TempDir = paste0(getwd(), "/Temp"),
  Save_Dir = paste0(getwd(), "/AgMERRA"),
  cores = 4,
  Years = c(1980, 2010),
  M.ORIGIN = "1900-01-01"
)
}
\arguments{
\item{DATA}{A data.table or data.frame containing decimal degree point locations as two numeric columns \code{Latitude} \code{Longitude}, an numeric field
\code{Buffer} that describes a radius of spatial uncertainty in meters, and a unique id field naming each location whose name is specified using the \code{ID} parameter.}

\item{ID}{A character vector of length one containing the column name for a unique id field naming each location in the dataset provided.}

\item{AgMERRA_dir}{A character vector of length one containing the path to the AgMERRA dataset}

\item{TempDir}{A character vector of length one containing the path to a temporary directory for AgMERRA processing. Files are saved here until analysis is completed, in case of interruptions.}

\item{Save_Dir}{A character vector of length one containing the path to the directory where the extracted AgMERRA data is to be saved. Set to NA if you do not want to save the returned dataset.}

\item{cores}{A numeric vector of length one specifying the number of logical cores to use for parallel processing (do not set to 0 or 1).  Set to NA to disable parallel processing. Default = 4.}

\item{Years}{A numeric vector of length two specifying the starting and end years for which AgMERRA should be extracted. Default = \code{c(1980,2010)}.}

\item{M.ORIGIN}{A character vector of length one in the form dddd-mm-dd specifying a date of origin to which dates are mapped to as the output \code{DayCount} field. Default value is \code{1900-01-01}}
}
\value{
A data.table of daily climate data including the fields:
\itemize{
\item \code{Year} = Year of observation (integer)
\item \code{Day}= Julian day of observation (integer)
\item \code{Temp.Mean} = mean temperature - C (numeric)
\item \code{Temp.Min} = minimum temperature - C (numeric)
\item \code{Temp.Max} = maximum temperature - C (numeric)
\item \code{Solar.Rad} = short wave solar radiation - MJ/m2/day (numeric)
\item \code{WindSpeed} = windspeed at 2 m above ground level -  m/s (numeric)
\item \code{RH.Max} = relative humidity at time of max temp - \% (numeric)
\item \code{Rain} = precipitation - mm (Numeric)
\item \code{Date} = date of observation (Date)
\item \code{DayCount} = days since date specified in \code{M.ORIGIN} parameter
}
}
\description{
This function extracts daily climate data from AgMERRA .nc4 files for a dataset containing point locations with a buffer of spatial uncertainty.
}
\details{
AgMERRA is the AgMIP climate forcing dataset based on the NASA Modern-Era Retrospective Analysis for Research and Applications (MERRA). AgMERRA
corrects to gridded temperature and precipitation, incorporates satellite precipitation, and derives solar radiation from NASA/GEWEX SRB to cover
the 1980-2010 period

You will need to download all AgMERRA files from \href{https://data.giss.nasa.gov/impacts/agmipcf/agmerra/}{this repository} a directory, then specify this
directory as the \code{AgMERRA_dir} argument.

Values are extracted and averaged for bounding boxes created from each unique point location and buffer in the data supplied.
}
