% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCHIRPS.R
\name{ExtractCHIRPS}
\alias{ExtractCHIRPS}
\title{Extract Climate Data from reformatted CHIRPS 2.0 dataset}
\usage{
ExtractCHIRPS(
  Data,
  ID,
  CHIRPS_dir,
  Save_Dir,
  YStart,
  YEnd,
  Round = 4,
  Origin = "1900-01-01"
)
}
\arguments{
\item{ID}{A character vector of length one containing the column name for a unique id field naming each location in the dataset provided.}

\item{CHIRPS_dir}{A character vector of length one containing the path to the directory where reformatted CHIRPS data is saved.}

\item{Save_Dir}{A character vector of length one containing the path to the directory where the extracted and compiled POWER data is to be saved. Set to NA if you do not want to save the returned dataset.}

\item{YStart}{A single integer value from 1983 onward indicating the start year for CHIRPS data extraction.}

\item{YEnd}{A single integer value from 1983 onward indicating the end year for CHIRPS data extraction.}

\item{Round}{A integer value indicating the number of decimal places to round the output \code{RAIN} field to. Default = \code{5}}

\item{Origin}{A character vector of length one in the form \code{dddd-mm-dd} specifying a date of origin to which dates are mapped to as the output
\code{DayCount} field. Default value is \code{1900-01-01}}

\item{DATA}{A data.table or data.frame containing decimal degree point locations as two numeric columns \code{Latitude} \code{Longitude}, an numeric field
\code{Buffer} that describes a radius of spatial uncertainty in meters, and a unique id field naming each location whose name is specified using the \code{ID} parameter.}
}
\value{
A data.table of daily precipitation data including the fields:
\itemize{
\item \code{RAIN} = precipitation - mm day-1 (numeric)
\item \code{R.Cells} = number of non NA cells contained in a locations bounding box
\item \code{R.Cells_0} = number of cells with 0 mm precipitation contained in a locations bounding box
\item \code{R.Cells_1} = number of cells with >= 1 mm precipitation contained in a locations bounding box
\item \code{R.Cells_5} = number of cells with >= 5 mm precipitation contained in a locations bounding box
\item \code{DAY} = julian day of observation (integer)
\item \code{LON} = longitude (decimal degrees)
\item \code{LAT} = latitude (decimal degrees)
\item \code{BUF} =  radius of spatial uncertainty by which location is buffer - m
\item \code{ID} = identity field as per \code{ID} parameter specified
\item \code{YEAR} = year of observation (integer)
\item \code{Date} = date of observation (Date)
\item \code{DayCount} = days since date specified in \code{M.ORIGIN} parameter
}
}
\description{
For a dataset containing point locations with a buffer of spatial uncertainty this function averages daily precipitation information from the CHIRPS 2.0 dataset.
}
\details{
To use this function you will first need to download ref{https://www.chc.ucsb.edu/data/chirps}{CHIRPS 2.0} data using the \code{DownloadCHIRPS} function
and then reformat this data using the \code{ReformatCHIRPS} function.
}
