% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFolder.R
\name{ExtractFolder}
\alias{ExtractFolder}
\title{Extract data from all tif files for a supplied vector layer}
\usage{
ExtractFolder(
  FileDirectory,
  MaxChunkSize,
  ExtractBy,
  ID.Target,
  ID.Target.Name = "variable",
  ID.ExtractBy,
  ID.ExtractBy.Name = NA,
  Function = mean,
  value.name = "value",
  NAValue = NA,
  Round = NA
)
}
\arguments{
\item{FileDirectory}{the location of the folder containing \code{.tif} rasters to be extracted, this folder should only contain \code{.tif} files.}

\item{MaxChunkSize}{the maximum number of rasters to be stacked when extracting data, consider your working RAM when making this decision.}

\item{ExtractBy}{a set of locations in \code{SpatVector} form.}

\item{ID.Target}{A character vector providing names identifying each raster in the \code{FileDirectory} folder. The vector must be the same length and order as the files in that folder.}

\item{ID.Target.Name}{the name of the raster identity field. Default = \code{variable}.}

\item{ID.ExtractBy}{A character vector providing names for each location in \code{ExtractBy}. The vector must be the same length and order as \code{ExtractBy}.}

\item{ID.ExtractBy.Name}{the name of the location field. Default = \code{ID.ExtractBy}.}

\item{Function}{the function to be applied to summarize values extracted for each location. If set to \code{NA} then no function is applied. Default = \code{mean}.}

\item{value.name}{the name of the value field. Default = \code{value}.}

\item{NAValue}{if NA values are encoded with specific value (e.g. -9999) use this field to state that value.}

\item{Round}{an integer indicating the number of decimal places to round the value field to.}
}
\value{
A data.table of daily precipitation data including the fields:
\itemize{
\item \code{value.name} = values contained within the supplied rasters
\item \code{ID.Target.Name} = identity of supplied rasters
\item \code{ID.ExtractBy.Name} = identity of extraction location
}
}
\description{
If parallel processing can be added then speeds can be increased. See https://github.com/rspatial/terra/issues/36 using pack and unpack functions to
make serializable SpatRaster and SpatVector objects.
With a small modification on (line 97) it should be possible to extract using a raster too.
}
