% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractMYOs.R
\name{ExtractMYOs}
\alias{ExtractMYOs}
\title{Multi-year Observations}
\usage{
ExtractMYOs(Data)
}
\arguments{
\item{Data}{A prepared (ERA) dataset, see the \code{ERA.Prepare} function.}

\item{Prac}{The column name of the variable in \code{Data} that describes the practice. Use this parameter to choose different levels of the practice hierarchy. Default = \code{PrName} (Practice).}

\item{Out.Codes}{A vector of outcome codes to consider in the analysis. Default = \code{101} (Crop Yield).}

\item{MinYear}{An integer value for the minimum length of a MYO sequence. Sequences with fewer growing season than this number are excluded from analysis. Default = \code{3}.}
}
\value{
A list of two data.tables \code{Risk.Means} & \code{Risk.Diff}, the difference between these is described under the Details section.
Output fields:
\itemize{
\item \code{UID} = a unique identifier based on the field \code{Outcome},\code{Practice},\code{Practice.Base},\code{Code},\code{ID},\code{Site.ID},\code{EU},\code{T.Descrip},\code{C.Descrip},\code{T.NI},\code{T.NO},\code{C.NI},\code{C.NO},\code{Tree},\code{Variety},\code{Diversity}, and \code{Rep}.
\item \code{N.Years}= the number of unique growing seasons reported for a value of \code{UID}.
\item \code{N.Obs} = the total number of observation for a value of \code{UID}.
}
}
\description{
This function labels multi-year observations with unique treatments, i.e. the same experiment, location, experimental vs control comparison, product and outcome.
}
\details{
A identifier \code{UID} is given to "unique" treatment x outcome x product combinations and the number of unique growing seasons \code{N.Years} and
total number of observations \code{N.Obs} is calculated for each \code{UID} is calculated. If \code{N.Obs} > \code{N.Years} this indicates that one or more ERA
practices varies, but this is not captured in the ERA dataset. In these unusual circumstances we suggest averaging data for the \code{UID} across seasons.
This situation may arise where there are \emph{levels} to an ERA practice such as different amounts of mulch application, but these were not captured in
the the Treatment description. The ERA team will be working to solve these issues in a future release.
}
