% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractPOWER.R
\name{ExtractPOWER}
\alias{ExtractPOWER}
\title{Extract Climate Data from NASA POWER agriclimatology dataset}
\usage{
ExtractPOWER(
  DATA,
  ID,
  Parameters = paste0("ALLSKY_SFC_SW_DWN,", "PRECTOT,", "PS,", "QV2M,", "RH2M,",
    "T2M,", "T2M_MAX,", "T2M_MIN,", "WS2M"),
  StartDate = 19830701,
  EndDate = 20181231,
  Save_Dir = NA,
  PowerSave = "POWER Downloads/",
  DELETE = F,
  M.ORIGIN = "1900-01-01",
  MaxBuffer = 240000
)
}
\arguments{
\item{DATA}{A data.table or data.frame containing decimal degree point locations as two numeric columns \code{Latitude} \code{Longitude}, an numeric field
\code{Buffer} that describes a radius of spatial uncertainty in meters, and a unique id field naming each location whose name is specified using the \code{ID} parameter.}

\item{ID}{A character vector of length one containing the column name for a unique id field naming each location in the dataset provided.}

\item{Parameters}{A character vector of length one contain the names of POWER variable to be downloaded. Excellent documentation for the POWER API using html web query can be found ref{https://power.larc.nasa.gov/docs/v1/}{here} or
ref{https://power.larc.nasa.gov/new/files/POWER_Data_v8_methodology.pdf}{here}. Default value = \code{paste0("ALLSKY_SFC_SW_DWN,", "PRECTOT,", "PS,","QV2M,","RH2M,","T2M,","T2M_MAX,","T2M_MIN,","WS2M")}}

\item{StartDate}{A numeric vector of length one containing the start date for POWER data extraction in the form \code{yyyymmdd}. The default value = \code{19830701} and this is the earliest date for which POWER data is available}

\item{EndDate}{A numeric vector of length one containing the start date for POWER data extraction in the form \code{yyyymmdd}. Default value = 20181231.}

\item{Save_Dir}{A character vector of length one containing the path to the directory where the extracted and compiled POWER data is to be saved. Set to NA if you do not want to save the returned dataset.}

\item{PowerSave}{A character vector of length one containing the path to a directory where download POWER data is saved.}

\item{DELETE}{Logical T/F. Delete downloaded site files once all sites are downloaded and the data has been processed and combined (default = FALSE)}

\item{M.ORIGIN}{A character vector of length one in the form \code{dddd-mm-dd} specifying a date of origin to which dates are mapped to as the output \code{DayCount} field. Default value is \code{1900-01-01}}

\item{MaxBuffer}{Maximum allowed buffer radius in m. Default and maximum value = \code{240000}.}
}
\value{
A data.table of daily climate data including the fields:
\itemize{
\item \code{Year} = Year of observation (integer)
\item \code{Day}= julian day of observation (integer)
\item \code{Solar.Rad} = insolation incident on a horizontal surface - MJ/m^2/day (numeric)
\item \code{Pressure} = surface pressure - kPa = (numeric)
\item \code{Specific.Humid.} = specific humidity at 2 meters - \% (numeric)
\item \code{Humid} = relative humidity at 2 meters - \% (numeric)
\item \code{Temp.Mean} = mean temperature - C (numeric)
\item \code{Temp.Min} = minimum temperature - C (numeric)
\item \code{Temp.Max} = maximum temperature - C (numeric)
\item \code{WindSpeed} = wind speed at 2 m -  m/s (numeric)
\item \code{Rain} = precipitation - mm day-1 (numeric)
\item \code{Date} = date of observation (Date)
\item \code{DayCount} = days since date specified in \code{M.ORIGIN} parameter
}
}
\description{
For a dataset containing point locations with a buffer of spatial uncertainty this function downloads and summarizes daily climate information from
the ref {NASA POWER}{https://power.larc.nasa.gov/} agroclimatology dataset at 0.5&deg resolution.
}
\details{
NASA POWER Solar parameters are derived from NASA's GEWEX/SRB release 3.0 archive (July 1, 1983 – Dec. 31, 2007) and
NASA’s CERES FLASHFlux project (Jan. 1, 2008 – to within about 7-days of real time). Meteorological parameters are derived from the NASA's GMAO
MERRA-2 assimilation model (Jan. 1, 1981 to within a few months of real time) plus GEOS-5.12.4 FP-IT (End of MERRA-2 to within several days of
real time).

Values are extracted and averaged for bounding boxes created from each unique point location and buffer in the data supplied.

Note older POWER Solar.Rad values are occasionally missing, these can be substituted using AgMERRA data and the \code{ReplaceSRAD} function.
}
