% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnrr_laj.R
\name{lnrr_laj}
\alias{lnrr_laj}
\title{lnrr_laj}
\usage{
lnrr_laj(m1, m2, cv1_2, cv2_2, n1, n2)
}
\arguments{
\item{m1}{Mean of treatment group 1}

\item{m2}{Mean of treatment group 2}

\item{cv1_2}{Coefficient of variation squared (CV^2) for treatment group 1}

\item{cv2_2}{Coefficient of variation squared (CV^2) for treatment group 2}

\item{n1}{Sample size for treatment group 1}

\item{n2}{Sample size for treatment group 2}
}
\description{
Calculates log response ratio based on Taylor expansion from Jajeunesse 2011.
}
\references{
Nakagawa, S., et al. (2022). "A robust and readily implementable method for the meta-analysis of response ratios with and without missing standard deviations." Ecol Lett. doi: 10.1111/ele.14144
}
