% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnrr_naka.R
\name{lnrr_naga}
\alias{lnrr_naga}
\title{lnrr_naga}
\usage{
lnrr_naga(m1, m2, n1, n2, n1_data, n2_data, CV1_data, CV2_data)
}
\arguments{
\item{m1}{Mean of treatment group 1}

\item{m2}{Mean of treatment group 2}

\item{n1}{Sample size for treatment group 1}

\item{n2}{Sample size for treatment group 2}

\item{n1_data}{Vector of sample sizes for treatment group 1 (for those observations with CV)}

\item{n2_data}{Vector of sample sizes for treatment group 2 (for those observations with CV)}

\item{CV1_data}{Vector of coefficient of variations for treatment group 1 (for those observations with CV)}

\item{CV2_data}{Coefficient of variation for treatment group 2 (for those observations with CV)}
}
\description{
Calculates log response ratio based on Taylor expansion from Lajeunesse 2011 as per equation 6 of Nagakawa 2022.
}
\references{
Nakagawa, S., et al. (2022). "A robust and readily implementable method for the meta-analysis of response ratios with and without missing standard deviations." Ecol Lett. doi: 10.1111/ele.14144
}
