% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchTreatments.R
\name{MatchTreatments}
\alias{MatchTreatments}
\title{Match Observations}
\usage{
MatchTreatments(Data, Match.Cols, Focal.Col)
}
\arguments{
\item{Data}{A prepared (ERA) dataset, see the \code{ERA.Prepare} function.}

\item{Match.Cols}{The column name of the variable in \code{Data} that describes the practice. Use this parameter to choose different levels of the practice hierarchy. Default = \code{PrName} (Practice).}

\item{Focal.Col}{A vector of outcome codes to consider in the analysis. Default = \code{101} (Crop Yield).}
}
\value{
A data.table with three columns appended::
\itemize{
\item \code{MID} = a unique identifier for each combination of the grouping variables (column names) supplied \code{Match.Cols}
\item \code{N.Focal}= the number of unique values of \code{Focal.Col} for each value of \code{MID}
\item \code{Focal.List} = unique values of \code{Focal.Col} concatenated with a \code{-} delim calculate for each value \code{MID}
}
}
\description{
This simple function will add a identfication field \code{MID} to the table supplied for unique combinations of the grouping columns supplied.
}
\details{
For example, if you wish to know which outcomes are shared between the same pairings of experimental and control treatments you could specify
\code{Match.Cols = c("Code","Site.ID","C.Descrip","T.Descrip","SubPrName","SubPrName.Base","M.Year")} and \code{Focal.Col = Out.SubInd} this would give a unique
\code{MID} to observations that have the same control and treatment practices, location, growing season and study. The \code{N.Focal} column of the output
dataset would tell you how many unique values of \code{Out.SubInd} are present in each group, and \code{Focal.List} would concatenate
these values.
}
