% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutCalc.R
\name{OutCalc}
\alias{OutCalc}
\title{Remove Extreme Outliers}
\usage{
OutCalc(Vals)
}
\arguments{
\item{Vals}{A vector of numeric values.}
}
\value{
A logical vector where \code{TRUE} indicate input values that are extreme outliers.
}
\description{
This simple function will detect extreme outliers (values above or below 3*interquartile range) in a supplied vector of numbers.
}
\details{
Outliers can be problematic in the analysis of ratio data when the control treatment (CT) tends to zero and the experimental treatment (ET) does not
Then the ET/CT ratio tends towards infinity which can unrealistically skew results’ distributions. This could occur when there are unequal impacts of
environmental stressors such as drought, heavy rainfall, pest or diseases between CT and ET.
}
