% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OutcomeCodes}
\alias{OutcomeCodes}
\title{Outcome Codes (Version 23.01.2023)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 115 rows and 19 columns.
}
\source{
ERA Project Team (2021)
}
\usage{
OutcomeCodes
}
\description{
\describe{The organizational hierarchy of outcomes considered in ERA and their descriptions.
\item{Code}{unique alpha numeric code}
\item{Pillar}{concepts in the highest level in the outcome hierarchy}
\item{Pillar.Code}{currently a duplicate of Pillar}
\item{Subpillar}{concepts in the intermediate level of the outcome hierarchy. Subpillar is nested under Pillar}
\item{Subpillar.Code}{short code corresponding to the Subpillar name for use in plotting}
\item{Indicator}{concepts in the intermediate level of the outcome hierarchy. Indicator is nested under Subpillar}
\item{Indicator.Code}{short code corresponding to the Indicator name for use in plotting}
\item{Subindicator}{concepts in the lowest and most detailed level of the outcome hierarchy. Subindicator is nested under Indicator}
\item{Subindicator.Short}{shorter name for the Subindicator}
\item{Subindicator.Code}{short code corresponding to the Subindicator name for use in plotting}
\item{Definition}{description of pthe Subindicator}
\item{Notes}{free text used to clarify the defition of the Subindicator}
\item{Example.units}{example of units that are collecting for that Subindicator}
\item{Original.Outcome}{currently NA}
\item{Negative.Values}{indicates whether the outcome can have negative values; N- No and Y-Yes}
\item{Sign}{indicates whether the outcome is positive or negative}
\item{TC.Ratio}{this is the ratio between the outcome for the treatment and control. TC = TreatmeantControl}
...
}
}
\keyword{datasets}
