% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETcalc.R
\name{PETcalc}
\alias{PETcalc}
\title{Calculate Penman-Monteith reference evapotranspiration}
\usage{
PETcalc(
  Tmin,
  Tmax,
  SRad,
  Wind,
  Pressure,
  Humid,
  Rain,
  YearDay,
  Latitude,
  Altitude
)
}
\arguments{
\item{Tmin}{a numeric vector of daily minimum air temperatures at 2m height (celcius)}

\item{Tmax}{a numeric vector of daily maximum air temperatures at 2m height (celcius)}

\item{SRad}{a numeric vector of daily net radiation at the crop surface,  daily solar radiation (MJ m-2 day-1)}

\item{Wind}{a numeric vector of  mean daily wind speed at 2 m height (ms-1)}

\item{Pressure}{a numeric vector of mean daily atmospheric pressure (kPa)}

\item{Humid}{a numeric vector of daily mean relative humidity (\%)}

\item{Rain}{a numeric vector of daily precipitation (mm). Can be set to NA if rainfall data is not supplied.}

\item{YearDay}{an integer vector julian day of the year}

\item{Latitude}{a numeric of latitude in decimal degrees}

\item{Altitude}{a numeric of altitude (meters above sea level) (m)}
}
\value{
A data.table with two fields:
\itemize{
\item \code{ETo} = reference evapotranspiration (mm)
\item \code{WBalance} = daily precipitation-ETo (mm) (if \code{Rain} argument supplied)
}
}
\description{
This function calculates reference evapotranspiration using the ref{FAO Penman-Monteith method}{http://www.fao.org/docrep/X0490E/x0490e06.htm}.
}
