% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PracticeCodes}
\alias{PracticeCodes}
\title{Practice Codes (version 31.01.2022)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 174 rows and 24 columns.
}
\source{
ERA Project Team (2021)
}
\usage{
PracticeCodes
}
\description{
\describe{The organizational hierarchy of improved farming practice concepts considered in ERA and their descriptions.
\item{Code}{unique alpha numeric code}
\item{Theme}{concepts in the highest level in the practice hierarchy}
\item{Theme.Code}{currently a duplicate of theme}
\item{Practice}{concepts in the intermediate level of the practice hierarchy. Practice is nested under Theme.}
\item{Practice.Code}{short code corresponding to the practice name for use in plotting}
\item{Subpractice}{concepts in the lowest and most detailed level of the practice hierarchy. Subpractice is nested under practice.}
\item{Subpractice.Code}{short code corresponding to the subpractice name for use in plotting}
\item{Subpractice.S}{shorter name for subpractice}
\item{Subpractice.Suffix}{suffix for the subpractice}
\item{Definition}{description of Subpractice}
\item{Notes}{free text used to clarify the defition of the practice}
...
}
}
\keyword{datasets}
