% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareChordData.R
\name{PrepareChordData}
\alias{PrepareChordData}
\title{Prepare Data for Chord Plot}
\usage{
PrepareChordData(Data, OutcomeCodes, from, to, Min.Chord.Perc = 1, ALevel)
}
\arguments{
\item{Data}{a \code{data.frame} or \code{data.table} of ERA data with column names standardized using the \code{StandColNames} function.}

\item{OutcomeCodes}{A vector of outcome codes to consider in the analysis. Default = \code{101} (Crop Yield).}

\item{from}{A character vector of length one defining the \code{from} column in the returned data to be supplied to the \link[circlize]{chordDiagram} function.}

\item{to}{A character vector of length one defining the \code{to} column in the returned data to be supplied to the \link[circlize]{chordDiagram} function.}

\item{Min.Chord.Perc}{A numeric vector of length one defining the minimum chord width to be considered, values less than \code{Perc.Neg} are combined and renamed to \code{Other}.}

\item{ALevel}{A character vector of length one containing a value from \code{ERAg::ERAConcepts$Agg.Levels$Agg}; this indicates the of spatial aggregation over which data are summarized.}
}
\value{
a \code{data.table} with row indicating from-to object pairs and columns:
*\code{from} = objects specified in \code{from} argument
*\code{to} = objects specified in \code{to} argument
*\code{value} = width of chord (e.g. number of unique instances of \code{ALevel} for from-to object pair)
*\code{yi} = strength of relationship for from-to object pair as response ratio
*\code{value} = strength of relationship for from-to object pair as response as vote count
}
\description{
This function prepares ERA data for use with the \code{ERAChordPlot} function. Chord values based on response ratios and vote counting are calculated for each
combination of values in the \code{from} x \code{to} fields specified aggregated spatially by the field specified in the \code{ALevel} function..
}
\details{
Column names in the ERA dataset (e.g. \code{ERA.Compiled}) must be standardized using the \code{StandColNames} function before use with this function.
}
