% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareXY.R
\name{PrepareXY}
\alias{PrepareXY}
\title{Prepare ERA Data for XY Plot}
\usage{
PrepareXY(
  Data,
  Comparison,
  CodeAxes = c("Yes", "Yes"),
  PLevel,
  OLevel,
  EULevel
)
}
\arguments{
\item{Data}{a \code{data.frame} or \code{data.table} of ERA data (e.g. \code{ERA.Compiled}).}

\item{Comparison}{a character vector of length one; accepted values: \code{PxO} practice vs. outcome,\code{PxEU} practice vs. product (experimental unit) or
\code{OxEU} outcome vs. product. These are the respective x and y axes to be used in plotting.}

\item{CodeAxes}{a character vector of length two; accepted values: \code{Yes} or \code{No}; default = \code{c("Yes","Yes")}; When set to \code{Yes} axes show coded names,
when \code{No} axes show full names. Position 1 of this parameters refers to the x-axis and position 2 the y-axis of plots.}

\item{PLevel}{character vector of length one; accepted values:  \code{T} theme, \code{P} practice or \code{S} subpractice. This parameter indicates the \strong{practice}
level for analysis.}

\item{OLevel}{character vector of length one; accepted values:  \code{P} pillar, \code{SP} subpillar, \code{I} indicator or \code{SI} subindicator. This parameter
indicates the \strong{outcome} level for analysis.}

\item{EULevel}{character vector of length one; accepted values:  \code{T} type, \code{S} subtype  or \code{P} product. This parameter indicates the \strong{product}
level for analysis. \emph{Note products are also referred to as experimental units (EU) in ERA}.}
}
\value{
\code{PrepareXY} returns a list with two elements:
\itemize{
\item \verb{[[Data]]} returns a \code{data.table} of the supplied dataset with focal columns renamed to \code{Axis1}, \code{Axis2} and \code{Group}.
\item \verb{[[Lab]]} returns a character vector of names corresponding to \code{Axis1}, \code{Axis2} and \code{Group}.
}
}
\description{
\code{PrepareXY} standardizes ERA column names for use in generalized plotting functions that require selection of two out of three of the higher order
ERA concepts of  \strong{practice}, \strong{outcome} and \strong{product}.
}
