% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatCHIRPS.R
\name{ReformatCHIRPS}
\alias{ReformatCHIRPS}
\title{Reformat CHIRPS 2.0 dataset}
\usage{
ReformatCHIRPS(CHIRPS_dir, Save_dir)
}
\arguments{
\item{CHIRPS_dir}{A character vector of length one containing the path to the directory where CHIRPS data should save. Default value = \verb{/CHIRPS}.}

\item{Save_dir}{A single integer value from 1983 onwards indicating the end year for which CHIRPS data should be downloaded. Default value = \code{2020}.}

\item{Parallel}{Memory requirements for parallel processing seem rather high best to keep as F, unless you have a large amount of RAM}
}
\description{
This function reformats the CHIRPS data downloaded by the \code{DownloadCHIRPS} function to be an array object from which data can be quickly extracted.
}
