% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceSRAD.R
\name{ReplaceSRAD}
\alias{ReplaceSRAD}
\title{Replace missing POWER solar radiation values}
\usage{
ReplaceSRAD(SRad, Rain, SeqLen)
}
\arguments{
\item{SRad}{A numeric vector of solar radiation values}

\item{Rain}{A numeric vector of precipitation values aligned with \code{Rad}}

\item{SeqLen}{An integer value that describes how many days to look either side of an NA value for substitution.}
}
\value{
A numeric vector of the supplied \code{SRad} values where NAs have been substituted with the neighboring value with the most similar rainfall amount to the
missing value (if non NA values are present).
}
\description{
This function replaces missing solar radiation values in the POWER dataset by searching for temporally adjacent values.
}
\details{
The solar radiation value of the day with the most similar rainfall to the NA value within +/- \code{SeqLen} days is substituted.
}
