% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOS_Dekad.R
\name{SOS_Dekad}
\alias{SOS_Dekad}
\title{Format date as yearly or monthly dekad (10-day period)}
\usage{
SOS_Dekad(Data, type = c("month", "year"), ...)
}
\arguments{
\item{type}{dekad of \code{month} (1:3) or \code{year} (1:36)}

\item{x}{class \code{Date} vector to convert to dekad}
}
\value{
integer dekad
}
\description{
\code{SOS_Dekad} transforms a class \code{Date} object to dekad of the year or month.
}
\details{
This function is taken from Melanie Bacou's https://github.com/tidyverse/lubridate/issues/617#issuecomment-521393447
}
\examples{
dekad(Sys.Date())
}
