% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOS_RSeason.R
\name{SOS_RSeason}
\alias{SOS_RSeason}
\title{Find rainy seasons}
\usage{
SOS_RSeason(RAIN, AI)
}
\arguments{
\item{RAIN}{a logical vector indicating where a starting condition of rainfall is \code{TRUE}}

\item{AI}{a logical vector indicating whether an aridity index threshold is \code{TRUE} or \code{FALSE}}
}
\value{
integer sequence
}
\description{
\code{SOS_RSeason} generates dekadal sequences where an initial rainfall condition (onset or SOS) has been met ending when an aridity condition is not met (end of season, EOS). It
takes two logical sequences of dekadal information \code{RAIN} and \code{AI} and determines rainy seasons from these.
}
\details{
The \code{RAIN} parameter indicates where a starting condition of rainfall is \code{TRUE}. The \code{AI} parameter indicates an ending condition
where the aridity index is \code{FALSE}. The growing season occurs for sequences starting when \code{AI} and \code{RAIN} are
\code{TRUE} and ending when \code{AI} is \code{FALSE.}

The typical WRSI onset date definition is a dekad with 25 mm followed by two dekads with a total of 20mm.
The typical WRSI end of season definition is when the long-term mean aridity index drops below 0.5 (i.e. mean potential evapotranspiration >= 2*mean rainfall).
}
