% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOS_SeasonPad.R
\name{SOS_SeasonPad}
\alias{SOS_SeasonPad}
\title{Pad season sequences}
\usage{
SOS_SeasonPad(Data, PadBack, PadForward)
}
\arguments{
\item{Data}{an integer vector of seasons (sequences of \code{1} and \code{2} separated by NAs)}

\item{PadBack}{an integer value specifying number of places to increase sequence backwards. This should not be greater than the minimum separation of sequences.}

\item{PadForward}{an integer value specifying number of places to increase sequence forwards. This should not be greater than the minimum separation of sequences.}
}
\value{
integer season
}
\description{
\code{SOS_SeasonPad} takes season sequences (i.e., values of 1 and 2 separated by NAs) and increases the sequence length in forward and backwards directions.
}
