% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOS_SeqMerge.R
\name{SOS_SeqMerge}
\alias{SOS_SeqMerge}
\title{Merge sequences}
\usage{
SOS_SeqMerge(Seq, AI, MaxGap, MinStartLen, MaxStartSep, ClipAI)
}
\arguments{
\item{Seq}{a vector of sequences split by NAs}

\item{AI}{a logical vector the same length as \code{Seq} indicating whether an aridity index threshold is \code{TRUE} or \code{FALSE}}

\item{MaxGap}{an integer value describing the maximum gap (number of NA values) allowed between non-NA values before the sequence breaks.}

\item{MinStartLen}{an integer value describing the minimum length of first sequence block, if the first sequence is too short and too separated from the next sequence it is removed. This is to remove false starts.}

\item{MaxStartSep}{an integer value describing the maximum separation of the first sequence block, if the first sequence is too short and too separated from the next sequence it is removed. This is to remove false starts.}

\item{ClipAI}{logical \code{T/F}, if \code{T} then \code{AI} values corresponding to the last non-NA value in \code{Seq} are all set to \code{F} and the sequence is halted by the last \code{F} value of AI.}
}
\value{
a vector where sequences are merged across NAs and set to the first value of the series. Leading and trailing NAs remain.
}
\description{
\code{SOS_SeqMerge} takes a sequence created by \code{SOS_RSeason} where multiple non-NA values are split by sequences of NA values and in the corresponding aridity index (\code{AI}) data identifies the start and end point of non-NA values split by NA sequences of length \verb{<=MaxGap}.
Values from the start to end points are set to the first numeric value in the sequence. Values outwith the start and end points are set to NA.
}
\details{
For example \verb{NA NA NA NA NA  2  2  2  2  2 NA  3  3  3 NA NA  4 NA NA} with \code{MaxVal=1} becomes \verb{NA NA NA NA NA  2  2  2  2  2  2  2  2  2 NA NA NA NA NA NA} and with \code{MaxVal=2}
becomes \verb{NA NA NA NA NA  2  2  2  2  2  2  2  2  2  2  2  2 NA NA}.

False starts can be dealt with using the \code{MinStartLen} and \code{MaxStartSep} arguments.

The \code{ClipAI} argument determines if all sequences ends are recalculated based on on the rules provided, or just the sequences that are
split by NAs. If \code{F} then cleaned/merged sequences are likely to have a different end dekad to sequences that are not split.
}
