% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOS_UniqueSeq.R
\name{SOS_UniqueSeq}
\alias{SOS_UniqueSeq}
\title{Give season sequences in a time-series unique values}
\usage{
SOS_UniqueSeq(Data)
}
\arguments{
\item{Data}{an integer vector of seasons (sequences of \code{1} and \code{2} separated by NAs)}
}
\value{
integer unique sequence
}
\description{
\code{SOS_UniqueSeq} takes season sequences (i.e., values of 1 and 2 separated by NAs) and gives each season a unique integer value.
}
