% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabCalc.R
\name{StabCalc}
\alias{StabCalc}
\title{Calculate Outcome Stability}
\usage{
StabCalc(
  Data,
  Do.Weight = T,
  Weight.by.Study = T,
  Rm.Out = T,
  Transform = T,
  DoRandom = T,
  Control = list(optimizer = "optim", optmethod = "Nelder-Mead", maxit = 10000),
  Responses = c("lnRR", "lnVR", "lnCVR"),
  Use.acr = F
)
}
\arguments{
\item{Data}{\emph{To be described}}

\item{Do.Weight}{\emph{To be described}}

\item{Weight.by.Study}{\emph{To be described}}

\item{Rm.Out}{\emph{To be described}}

\item{Transform}{\emph{To be described}}

\item{Control}{\emph{To be described}}

\item{Responses}{\emph{To be described}}

\item{Use.acr}{logical T/F. If T scale-adjusted coefficient of variation, acv, is substituted for the coefficient of variation (cv).}
}
\value{
\code{StabCalc} returns a \code{data.table} \emph{to be described}
}
\description{
\code{StabCalc} calculates relative and absolute outcome stability for MYOs based on the methods of \href{https://doi.org/10.1038/s41467-018-05956-1}{Knapp \emph{el al.} (2018)}
}
