% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabCalc2.R
\name{StabCalc2}
\alias{StabCalc2}
\title{Calculate Outcome Stability}
\usage{
StabCalc2(
  Data,
  Do.Weight = T,
  Weight.by.Study = T,
  Rm.Out = T,
  Transform = T,
  DoRandom = T,
  Control = list(optimizer = "optim", optmethod = "Nelder-Mead", maxit = 10000),
  Responses = c("lnRR", "lnVR", "lnCVR"),
  Use.acv = F
)
}
\arguments{
\item{Data}{A data.table output by the \code{ERAg::PrepareStabData} function}

\item{Do.Weight}{logical, if \code{TRUE} coefficient estimates are weighted acccording to the supplied weightings in the \code{Data} object supplied(default = T)}

\item{Weight.by.Study}{logical, depreciated (default = \code{T})}

\item{Rm.Out}{logical, if \code{TRUE} extreme outliers are removed withing each Practice x Outcome combination as per the method detailed in \code{ERAg::OutCalc} (default = \code{T})}

\item{Transform}{logical, if \code{TRUE} back-transformed coefficient estimates and confidence intervals are appended to outputs  (default = \code{T})}

\item{Control}{list, optional list of control values for the \code{rma.mv} estimation algorithms. If unspecified, default values are defined inside the function (default = \code{list(optimizer="optim",optmethod="Nelder-Mead",maxit=10000)})}

\item{Responses}{character vector, this argument is depreciated do edit (default=\code{c("lnRR","lnVR","lnCVR")})}

\item{Use.acv}{logical T/F. If T scale-adjusted coefficient of variation, acv, is substituted for the coefficient of variation (cv).}
}
\value{
\code{StabCalc2} returns a \code{data.table}
Output fields:
\itemize{
\item 
\itemize{
\item To be described*
}
}
}
\description{
\code{StabCalc2}runs splits the \code{data.table} output by \code{PrepareStabData} into a list using the \code{Outcome} & \code{Practice} fields then applies \code{StabCalc}
to each element of the list.
}
