% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandColNames.R
\name{StandColNames}
\alias{StandColNames}
\title{Rename Concept Columns}
\usage{
StandColNames(Data, PLevel, OLevel, EULevel)
}
\arguments{
\item{Data}{a \code{data.frame} or \code{data.table} of ERA data (e.g. \code{ERA.Compiled}).}

\item{PLevel}{character vector of length one; accepted values:  \code{T} theme, \code{P} practice or \code{S} subpractice. This parameter indicates the \strong{practice}
level for analysis.}

\item{OLevel}{character vector of length one; accepted values:  \code{P} pillar, \code{SP} subpillar, \code{I} indicator or \code{SI} subindicator. This parameter
indicates the \strong{outcome} level for analysis.}

\item{EULevel}{character vector of length one; accepted values:  \code{T} type, \code{S} subtype  or \code{P} product. This parameter indicates the \strong{product}
level for analysis. \emph{Note products are also referred to as experimental units (EU) in ERA}.}
}
\value{
\code{StandColNames} returns a \code{data.table} of the supplied dataset with focal columns renamed to \code{Practice}, \code{Practice.Code}, \code{Practice.Base},
\code{Practice.Base.Code}, \code{Outcome}, \code{Outcome.Code}, \code{Product}, and \verb{Product Code}.
}
\description{
\code{StandColNames} renames ERA columns as per user selected values for organizational levels of Practice, Outcome and Product (EU) concepts.
}
\details{
The columns in \code{Data} corresponding to the organizational levels selected are renamed to \code{Practice}, \code{Outcome} and \code{Product}. This standardization of
column naming allows generalization of column naming for use with other functions in the \code{ERA} package  that assume these columns are present. This is
particularly useful for the creation of interactive apps where the user can choose to explore the data at different levels within the ERA concept hierarchy.

See the \code{CreateConceptLevels} function how parameter codes are translated to ERA column names.
}
