% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_lnrr_laj.R
\name{v_lnrr_laj}
\alias{v_lnrr_laj}
\title{v_lnrr_laj}
\usage{
v_lnrr_laj(cv1_2, cv2_2, n1, n2)
}
\arguments{
\item{cv1_2}{Coefficient of variation squared (CV^2) for treatment group 1}

\item{cv2_2}{Coefficient of variation squared (CV^2) for treatment group 2}

\item{n1}{Sample size for treatment group 1}

\item{n2}{Sample size for treatment group 2}
}
\description{
Calculates the sampling variance for log response ratio based on second order Taylor expansion proposed by Lajeunesse 2011
}
\references{
Nakagawa, S., et al. (2022). "A robust and readily implementable method for the meta-analysis of response ratios with and without missing standard deviations." Ecol Lett. doi: 10.1111/ele.14144
}
