% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_lnrr_naka.R
\name{v_lnrr_naka}
\alias{v_lnrr_naka}
\title{v_lnrr_naka
Calculates the sampling variance for log response ratio based on second order Taylor expansion proposed by Lajeunesse 2011 as per equation 7 of Nagakawa 2022}
\usage{
v_lnrr_naka(n1, n2, n1_data, n2_data, CV1_data, CV2_data)
}
\arguments{
\item{n1}{Sample size for treatment group 1}

\item{n2}{Sample size for treatment group 2}

\item{n1_data}{Vector of sample sizes for treatment group 1 (for those observations with CV)}

\item{n2_data}{Vector of sample sizes for treatment group 2 (for those observations with CV)}

\item{CV1_data}{Vector of coefficient of variations for treatment group 1 (for those observations with CV)}

\item{CV2_data}{Coefficient of variation for treatment group 2 (for those observations with CV)}
}
\description{
v_lnrr_naka
Calculates the sampling variance for log response ratio based on second order Taylor expansion proposed by Lajeunesse 2011 as per equation 7 of Nagakawa 2022
}
\references{
Nakagawa, S., et al. (2022). "A robust and readily implementable method for the meta-analysis of response ratios with and without missing standard deviations." Ecol Lett. doi: 10.1111/ele.14144
}
