% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyNegVals.R
\name{ClassifyNegVals}
\alias{ClassifyNegVals}
\title{Classify outcomes with negative numbers that are incompatible with ratio approaches}
\usage{
ClassifyNegVals(
  Data,
  OCode,
  Thresholds = c(0.7, 0.95, 1.05, 1.3),
  Vals = c(`0` = 0, `++` = 2, `+` = 1, `-` = -1, `--` = -2, `+-` = 1, `-+` = -1, `+0` =
    1, `0+` = -1, `-0` = -1, `0-` = 1),
  Invert2xNeg = T
)
}
\arguments{
\item{Data}{An ERA data.table (e.g. \code{ERAg::ERA.Compiled}).}

\item{OCode}{An ERA outcome code (see \code{ERAg::OutcomeCodes[,Codes]}.}

\item{Thresholds}{A numeric vector of length four corresponding to classification thresholds for worst, poor, good & best outcomes. If:
\itemize{
\item \code{MeanT/MeanC<Threshold[1]} then Class = \verb{--}
\item \verb{Threshold[1]<=MeanT/MeanC<Threshold[2]} then Class = \code{-}
\item \verb{Threshold[2]<=MeanT/MeanC<Threshold[3]} then Class = \code{0}
\item \verb{Threshold[3]<=MeanT/MeanC<Threshold[4]} then Class = \code{+}
\item \verb{MeanT/MeanC>=Threshold[4]`` then Class = }++`
}}

\item{Vals}{A numeric vector of length 11 proving numeric scores for each outcome classification:
\itemize{
\item \code{Vals[1]}: Class \code{0} = \code{MeanT/MeanC<Threshold[1]}
\item \code{Vals[2]}: Class \verb{++} = \verb{Threshold[1]<=MeanT/MeanC<Threshold[2]}
\item \code{Vals[3]}: Class \code{+}= \verb{Threshold[2]<=MeanT/MeanC<Threshold[3]}
\item \code{Vals[4]}: Class \code{-} = \verb{Threshold[3]<=MeanT/MeanC<Threshold[4]}
\item \code{Vals[5]}: Class \verb{--} = \code{MeanT/MeanC>=Threshold[4]}
\item \code{Vals[6]}: Class \verb{+-} = \code{MeanT>0 & MeanC<0}
\item \code{Vals[7]}: Class \verb{-+} =\code{ MeanT<0 & MeanC>0}
\item \code{Vals[8]}: Class \code{+0} = \code{MeanT>0 & MeanC==0}
\item \code{Vals[9]}: Class \verb{0+} = \code{MeanT==0 & MeanC>0}
\item \code{Vals[10]}: Class \code{-0} = \code{MeanT<0 & MeanC==0}
\item \code{Vals[11]}: Class \verb{0-} = \code{MeanT==0 & MeanC<0}
}}

\item{Invert2xNeg}{Logical \code{T/F}. Where MeanC<0 & MeanT<0 invert the calculation of yi to \code{log(MeanC/MeanT)}}
}
\value{
ClassifyNegVals returns the input Data subset to the outcome provided in the \code{OCode}argument and with the additional columns:
\itemize{
\item \code{Class} a character vector of class names
\item \code{Class.Val} a numeric vector of class values
}
}
\description{
For a the supplied outcome argument \code{OCode} this function classifies the difference between \code{MeanC} and \code{MeanT} into character and numeric columns based on the
\code{Threshold} and \code{Vals} arguments. These columns are appended to the input dataset that has been subset to the outcome of interest.
}
\details{
When generating ratios the function will invert outcomes where \code{MeanT} and \code{MeanC} are both negative unless the \code{Invert2xNeg} argument is set to \code{FALSE}.

Where MeanC is NA and MeanT is populated this function will base classfications on the MeanT value only.
}
