% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeriveOutcomes.R
\name{DeriveOutcomes}
\alias{DeriveOutcomes}
\title{Derive ERA outcomes from existing data}
\usage{
DeriveOutcomes(Data = ERA.Compiled, RmPartial = T, DoBCR_VC = T, DoWUE = T)
}
\arguments{
\item{Data}{An ERA data.table (e.g. \code{ERAg::ERA.Compiled}).}

\item{RmPartial}{Logical \code{T/F}. If \code{T} partial economic outcomes are excluded from the process.}

\item{DoBCR_VC}{Logical \code{T/F}. If \code{T} benefit cost ratios calculated using variable costs are also included in calculations as per description points 5) & 6).}

\item{DoWUE}{Logical \code{T/F}. If \code{T} water use efficiency outcome derived from yield and TSP (total seasonal precipitation) data}
}
\value{
ERAWeights returns the input \code{data.table} with additional rows appended for derived outcomes.
}
\description{
This function identifies exactly colocated outcomes (same experiment, treatment, location and time) ERA data that can be combined
to create additional outcome observations, generates these additional outcomes and appends them to the supplied Data.
}
\details{
The function currently generates additional outcomes for:
\enumerate{
\item Net Returns = gross returns - total costs
\item Gross Margin = gross returns - variable costs
\item Benefit Cost Ratio (GRTC) =  gross returns/total costs
\item Benefit Cost Ratio (NRTC) =  net returns/total costs
\item Benefit Cost Ratio (GRVC) =  gross returns/variable costs
\item Benefit Cost Ratio (NRVC) =  net returns/variable costs
}

Derived net returns are appended to \code{Data} before benefit cost ratio outcomes are derived.

Note that many of the outcomes derived from the data can be negative therefore be careful not to exclude these outcomes when applying the \code{ERAg::PrepareERA}
function (the \code{Perc.Neg} parameter could be adjusted to retain negative outcomes).
}
