% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadCHIRPS.R
\name{DownloadCHIRPS}
\alias{DownloadCHIRPS}
\title{Download CHIRPS 2.0 dataset}
\usage{
DownloadCHIRPS(StartYear = 1980, EndYear = 2021, EndDay = 365, SaveDir, quiet)
}
\arguments{
\item{StartYear}{A single integer value from 1983 onward indicating the start year for data download. Default value = \code{1981}.}

\item{EndYear}{A single integer value from 1983 onward indicating the end year for data download. Default value = \code{2021}.}

\item{EndDay}{A single integer value for the last day, in julian format, of available data for \code{EndYear}. Default value = \code{365}.}

\item{SaveDir}{A character vector of length one containing the path to the directory where CHIRPS data should save. Default value = \verb{/CHIRPS}.}

\item{quiet}{logical \code{T/F}, if \code{T} progress bar is suppressed}
}
\description{
This function download the CHIRPS 2.0 dataset at 0.05 degree resolution from the repository https://data.chc.ucsb.edu/products/CHIRPS-2.0/africa_daily/tifs/p05/1981/ to
a local directory specied by the \code{SaveDir} argument.
}
