% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstPDayData.R
\name{EstPDayData}
\alias{EstPDayData}
\title{Estimate missing planting dates from existing values in ERA}
\usage{
EstPDayData(DATA)
}
\arguments{
\item{DATA}{An ERA dataset (e.g. \code{ERA.Compiled})}
}
\value{
A list of two elements is returned:
\enumerate{
\item \verb{[[DATA]]} a data.table where the following fields are appended to the input dataset:
\code{Data.PS.Date} = planting start date estimated from the dataset provided (Date)
\code{Data.PE.Date} = planting end date estimated from the dataset provided (Date)
\code{Data.Date.Acc} = a concantenation of the number of decimals places used in latitude/longitude matching and whether matching was conducted using
\code{Product} (\code{P}) or \code{Product.Subtype} (\code{P.Sub}), for example \code{3-P.Sub}.
\item \verb{[[Season.Issues]]} a data.table containing Latitude + Longitude combinations where a mixture of season = 1 or 2 and season = NA are reported suggesting
potential issues with consistency of temporal recording at these spatial coordinates.
}
}
\description{
We may wish to substitute NA values of planting date in ERA with estimates derived from  same \code{Product/Product.Subtype}, year and
growing season which are spatially nearby, the \code{EstPDayData} functions generates these estimates.
}
\details{
The function searches iteratively within five levels of increasing distance corresponding to latitude and longitude recorded from 5 to 1 decimal
places. If there is more than one planting date value for a \code{Product} x \code{M.Year.Start} x \code{M.Year.End} x \code{Season.Start} x \code{Season.End}
combination then values are averaged. If no corresponding planting dates are found matching on product then matching is attempted
on \code{Product.Subtype} (e.g. cereals or legumes).

If Latitude + Longitude combinations have a mixture of season = 1 or 2 and season = NA values this suggests a potential issue
with the consistency of temporal recording at the spatial coordinates. Observations with such issues are excluded from the analysis and flagged in
the output \verb{[[Season.Issues]]} list object.

Observation with irrigation in the control and treatment are removed from consideration as these will not have realistic planting dates for rainfed areas.
}
