% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstSLenData.R
\name{EstSLenData}
\alias{EstSLenData}
\title{Estimate missing season lengths from existing values in ERA}
\usage{
EstSLenData(DATA)
}
\arguments{
\item{DATA}{An ERA dataset (e.g. \code{ERA.Compiled})}
}
\value{
The following fields are appended to the input dataset:
\code{Data.SLen} = season length date estimated from the dataset provided (days)
\code{Data.SLen.Acc} = a concantenation of the number of decimals places used in latitude/longitude matching and whether matching was conducted using
\code{Product} (\code{P.Sub}) or \code{Product.Subtype} (\code{P}), for example \code{3-P.Sub}.
}
\description{
We may wish to substitute NA values of season length (i.e. harvest date is unknown) in ERA with estimates derived from  same \code{Product/Product.Subtype}, year and
growing season which are spatially nearby, the \code{EstPDayData} functions generates these estimates.
}
\details{
Season length is calculated as the mean difference between mean harvest date (midpoint of \code{Harvest.Start} and \code{Harvest.End} Date columns) and mean planting date
(midpoint of \code{Plant.Start} and \code{Plant.End} Date columns).

The function searches iteratively within five levels of increasing distance corresponding to latitude and longitude recorded from 5 to 1 decimal
places. If there is more than one season length value for a \code{Product} x season (\code{M.Year}) combination then values are averaged. If no corresponding
season lengths are found matching on product then matching is attempted on \code{Product.Subtype} (e.g. cereals or legumes).
}
