% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractPOWER.R
\name{ExtractPOWER}
\alias{ExtractPOWER}
\title{Extract Climate Data from NASA POWER agriclimatology dataset}
\usage{
ExtractPOWER(
  Data,
  ID,
  Parameters = c("ALLSKY_SFC_SW_DWN", "PRECTOT", "PS", "QV2M", "RH2M", "T2M",
    "T2M_MAX", "T2M_MIN", "WS2M"),
  Rename = c("Solar.Rad", "Rain", "Pressure", "Humid", "Temp.Min", "Temp.Max",
    "Temp.Mean", "WindSpeed", "Specific.Humid"),
  StartDate = "1983-07-01",
  EndDate = "2021-12-31",
  Save_Dir = "POWER",
  PowerSave = "POWER/Downloads",
  Delete = F,
  MaxBuffer = 240000,
  AddDate = T,
  AddDayCount = T,
  Origin = "1900-01-01",
  Round = 2,
  Quiet = T
)
}
\arguments{
\item{Data}{A data.table or data.frame containing numeric fields for  \code{Latitude} and \code{Longitude} (decimal degrees), \code{Altitude} (m), and \code{Buffer}  (m). Buffer describes a radius of spatial uncertainty.
A character field naming each location is also required, the name of this column must be specified using the \code{ID} parameter.}

\item{ID}{A character vector of length one containing the column name for location name field in \code{Data}.}

\item{Parameters}{A character vector containing the names of POWER variables to be downloaded.}

\item{Rename}{A character vector of new variable matching the length of \code{Parameters}. Set to \code{NA} to retain orginal parameter names.}

\item{StartDate}{A numeric vector of length one containing the start date for POWER data extraction in the form \code{yyyy-mm-dd}. Default value = \code{1983-07-01} and this is the earliest date for which POWER data is available}

\item{EndDate}{A numeric vector of length one containing the start date for POWER data extraction in the form \code{yyyy-mm-dd}.}

\item{Save_Dir}{A character vector of length one containing the path to the directory where the extracted and compiled POWER data is to be saved. Set to NA if you do not want to save the returned dataset.}

\item{PowerSave}{A character vector of length one containing the path to a directory where download POWER data is saved.}

\item{Delete}{Logical \code{T/F}. Delete downloaded site files once all sites are downloaded and the data has been processed and combined (default = FALSE)}

\item{MaxBuffer}{Maximum allowed buffer radius in m. Default and maximum value = \code{240000}.}

\item{AddDate}{Logical \code{T/F}.If \code{T} then a date field is added to the output data.table.}

\item{AddDayCount}{Logical T/F. If \code{T} then an integer julian day field is added to the output data.table.}

\item{Origin}{A character vector of length one in the form \code{yyyy-mm-dd} specifying a date of origin to which dates are mapped to as the output \code{DayCount} field. Default value is \code{1900-01-01}}

\item{Round}{An integer value specifying the number of decimal places to round variable values to.}

\item{Quiet}{Logical T/F. Sets the \code{quiet} argument of the \code{download.file} function, if \code{T} a progress bar is displayed.}
}
\value{
A data.table of daily climate data, if default paramters are supplied then output fields are:
\itemize{
\item \code{Year} = Year of observation (integer)
\item \code{Day}= julian day of observation (integer)
\item \code{(ID)} = location name (character)
\item \code{Altitude} = altitude of site - m (numeric)
\item \code{Solar.Rad} = insolation incident on a horizontal surface - MJ/m^2/day (numeric)
\item \code{Pressure} = surface pressure - kPa = (numeric)
\item \code{Specific.Humid.} = specific humidity at 2 meters - \% (numeric)
\item \code{Humid} = relative humidity at 2 meters - \% (numeric)
\item \code{Temp.Mean} = mean temperature - C (numeric)
\item \code{Temp.Min} = minimum temperature - C (numeric)
\item \code{Temp.Max} = maximum temperature - C (numeric)
\item \code{WindSpeed} = wind speed at 2 m -  m/s (numeric)
\item \code{Pressure.Adjusted}= surface pressure adjusted to  - kPa = (numeric)
\item \code{Rain} = precipitation - mm day-1 (numeric)
\item \code{Date} = date of observation (Date)
\item \code{DayCount} = days since date specified in \code{Origin} parameter
}
}
\description{
For a dataset containing point locations with a buffer of spatial uncertainty this function downloads and summarizes daily climate information from
the  \href{NASA POWER}{https://power.larc.nasa.gov/} agroclimatology dataset at 0.5&deg resolution.
}
\details{
NASA POWER Solar parameters are derived from NASA's GEWEX/SRB release 3.0 archive (July 1, 1983 – Dec. 31, 2007) and
NASA’s CERES FLASHFlux project (Jan. 1, 2008 – to within about 7-days of real time). Meteorological parameters are derived from the NASA's GMAO
MERRA-2 assimilation model (Jan. 1, 1981 to within a few months of real time) plus GEOS-5.12.4 FP-IT (End of MERRA-2 to within several days of
real time).

Values are extracted and averaged for bounding boxes created from each unique point location and buffer in the data supplied.

Parameter descriptions are \href{here}{https://power.larc.nasa.gov/#resources}, details on API query formulation are \href{here}{https://power.larc.nasa.gov/docs/services/api/temporal/daily/} or
\href{here}{https://power.larc.nasa.gov/api/pages/?urls.primaryName=Daily}

Note older POWER Solar.Rad values are occasionally missing, these can be substituted using AgMERRA data and the \code{ReplaceSRAD} function.
}
