% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchTreatments.R
\name{MatchTreatments}
\alias{MatchTreatments}
\title{Match Observations}
\usage{
MatchTreatments(Data, Match.Cols, Focal.Col)
}
\arguments{
\item{Data}{An ERA dataset (e.g. \code{ERAg::ERACompiled})}

\item{Match.Cols}{A character vector of column names that must all contain identical values for rows in the dataset to match (i.e. share the same \code{MID})}

\item{Focal.Col}{A character vector of a single column name; the number of unique values of \code{Focal.Col} for each combination of value of \code{Match.Cols} values
(\code{MID}) are returned along with a concatenated string of the values.}
}
\value{
A data.table with three columns appended::
\itemize{
\item \code{MID} = a unique identifier for each combination of the grouping variables (column names) supplied \code{Match.Cols}
\item \code{N.Focal}= the number of unique values of \code{Focal.Col} for each value of \code{MID}
\item \code{Focal.List} = unique values of \code{Focal.Col} concatenated with a \code{-} delim calculate for each value \code{MID}
}
}
\description{
This simple function will add a identification field \code{MID} to the table supplied for unique combinations of the grouping columns supplied.
}
\details{
For example, if you wish to know which outcomes are shared between the same pairings of experimental and control treatments you would specify
\code{Match.Cols = c("Code","Site.ID","C.Descrip","T.Descrip","SubPrName","SubPrName.Base","M.Year")} and \code{Focal.Col = Out.SubInd} this gives a unique
\code{MID} to observations that have the same control and treatment practices, location, growing season and study. The \code{N.Focal} column of the output
dataset tells you how many unique values of \code{Out.SubInd} are present in each group, and  the \code{Focal.List} field in the output data.table
concatenates the unique values in alphabetical order with a - delimiter.
}
