% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutCalc.R
\name{OutCalc}
\alias{OutCalc}
\title{Remove Extreme Outliers}
\usage{
OutCalc(Vals)
}
\arguments{
\item{Vals}{A vector of numeric values.}
}
\value{
A logical vector where \code{TRUE} indicate input values that are extreme outliers.
}
\description{
This simple function will detect extreme outliers (values above or below 3*interquartile range) in a supplied vector of numbers.
}
