% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pbuffer.R
\name{Pbuffer}
\alias{Pbuffer}
\title{Buffer Points}
\usage{
Pbuffer(Data, ID = NA, Projected = F)
}
\arguments{
\item{Data}{A data.table or data.frame with decimal degree point locations in columns \code{Latitude} and \code{Longitude}, and a numeric column \code{Buffer} which indicates the buffer radius
to buffer each point in meters. NA values are not permitted and are filtered from the dataset.}

\item{ID}{The column name of any grouping variables used to split point x buffer locations. Default = NA.}

\item{Projected}{\verb{Logical TRUE/FALSE}. If \code{TRUE} the output object has a projected CRS \code{+proj=merc +lon_0=0 +k=1 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs},
if \code{FALSE} it has a geographic CRS \code{epsg:4326}.}
}
\value{
Pbuffer returns circular buffers in an object of class \code{SpatialPolygons}.
}
\description{
Creates circular buffers from a table containing point locations and buffer radii. The input table requires \code{Latitude} and \code{Longitude} columns in
decimal degrees and a numeric column named \code{Buffer} which indicates the buffer radius in meters.
}
\details{
An optional \code{ID} field can be specified if there is need to separate any points with identical locations and buffers.
}
