% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareStabData.R
\name{PrepareStabData}
\alias{PrepareStabData}
\title{Prepare Data for Outcome Stability Analysis}
\usage{
PrepareStabData(Data, OutCodes = 101)
}
\arguments{
\item{Data}{a \code{data.frame} or \code{data.table} of ERA data with column names standardized using the \code{StandColNames} function.}

\item{OutCodes}{A vector of outcome codes to consider in the analysis. Default = \code{101} (Crop Yield).}
}
\value{
a \code{data.table} of the supplied data with appended columns:
\itemize{
\item \emph{column descriptions to be added}
}
}
\description{
This function prepares data for use with the \code{StabCalc} function
}
