% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabCalc2.R
\name{StabCalc2}
\alias{StabCalc2}
\title{Calculate Outcome Stability for Each Element of Practice x Outcome}
\usage{
StabCalc2(
  Data,
  Do.Weight = TRUE,
  Weight.by.Study = TRUE,
  Rm.Out = TRUE,
  Transform = TRUE,
  DoRandom = TRUE,
  Control = list(optimizer = "optim", optmethod = "Nelder-Mead", maxit = 10000),
  Responses = c("lnRR", "lnVR", "lnCVR"),
  Use.acv = FALSE,
  Inc.Prod = FALSE,
  Cores = 1
)
}
\arguments{
\item{Data}{A \code{data.table} generated by the \code{ERAg::PrepareStabData} function.}

\item{Do.Weight}{Logical. If \code{TRUE}, coefficient estimates are weighted according to the weightings in the \code{Data} object (default = TRUE).}

\item{Weight.by.Study}{Logical. This parameter is deprecated (default = TRUE).}

\item{Rm.Out}{Logical. If \code{TRUE}, extreme outliers are removed within each Practice x Outcome combination (default = TRUE).}

\item{Transform}{Logical. If \code{TRUE}, back-transformed coefficient estimates and confidence intervals are appended to outputs (default = TRUE).}

\item{DoRandom}{Logical. If \code{TRUE}, random effects models are included (default = TRUE).}

\item{Control}{List of control values for the \code{rma.mv} estimation algorithm. Defaults to \code{list(optimizer="optim", optmethod="Nelder-Mead", maxit=10000)}.}

\item{Responses}{Character vector of response variables (default = \code{c("lnRR","lnVR","lnCVR")}).}

\item{Use.acv}{Logical. If \code{TRUE}, uses the scale-adjusted coefficient of variation (acv) instead of the coefficient of variation (cv) (default = FALSE).}

\item{Inc.Prod}{Logical. If \code{TRUE}, analysis is conducted at the Crop x Outcome x Practice level (default = FALSE).}

\item{Cores}{Integer. Number of cores to use for parallel processing. If \code{Cores = 1}, runs sequentially (default = 1).}
}
\value{
A list containing:
\enumerate{
\item \verb{[[StabStats]]}: Results from \code{ERAg::StabCalc} for each Practice x Outcome (and EU if \code{Inc.Prod = TRUE}).
\item \verb{[[StabStats.Test]]}: Equivalent of the \verb{[[Tests]]} output from \code{ERAg::StabCalc}.
\item \verb{[[StabStats.Tab]]}: Equivalent of the \verb{[[Coefs]]} output from \code{ERAg::StabCalc}.
\item \verb{[[StabStats.Test2]]}: Equivalent of the \verb{[[Tests2]]} output from \code{ERAg::StabCalc}.
\item \verb{[[StabList]]}: A list of the input \code{Data}, split by Practice and Outcome (and EU if \code{Inc.Prod = TRUE}).
}
}
\description{
This function splits the \code{data.table} output by \code{PrepareStabData} into a list using the \code{Outcome} and \code{Practice} fields,
then applies \code{ERAg::StabCalc} to each element of the list. The results are concatenated and returned in a list.
}
