% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StabCalc2.R
\name{StabCalc2}
\alias{StabCalc2}
\title{Calculate Outcome Stability}
\usage{
StabCalc2(
  Data,
  Do.Weight = T,
  Weight.by.Study = T,
  Rm.Out = T,
  Transform = T,
  DoRandom = T,
  Control = list(optimizer = "optim", optmethod = "Nelder-Mead", maxit = 10000),
  Responses = c("lnRR", "lnVR", "lnCVR"),
  Use.acv = F
)
}
\arguments{
\item{Data}{\emph{To be described}}

\item{Do.Weight}{\emph{To be described}}

\item{Weight.by.Study}{\emph{To be described}}

\item{Rm.Out}{\emph{To be described}}

\item{Transform}{\emph{To be described}}

\item{Control}{\emph{To be described}}

\item{Responses}{\emph{To be described}}

\item{Use.acv}{logical T/F. If T scale-adjusted coefficient of variation, acv, is substituted for the coefficient of variation (cv).}
}
\value{
\code{StabCalc2} returns a \code{data.table}
Output fields:
\itemize{
\item 
\itemize{
\item To be described*
}
}
}
\description{
\code{StabCalc2}runs splits the \code{data.table} output by \code{PrepareStabData} into a list using the \code{Outcome} & \code{Practice} fields then applies \code{StabCalc}
to each element of the list.
}
